refactor(app): move the last seven QuantickApp groups into app/ sidecars
`app.rs` was 5,388 production lines and still held every group the first
three cuts left behind. Each one already had a sibling-module shape waiting,
so each gets its own address:

- `control_host.rs` — the accessors the control plane reads the cockpit
  through, the two dispatch points, and the focused-pane helpers.
- `tabs.rs` — opening, closing and cycling tabs, plus the two free functions
  a saved workspace's bar rule is read with.
- `toolbar_wiring.rs` — the window side of the toolbar.
- `menu_bar.rs` — the menu bar, the tab strip, and the shortcut constants
  they mirror.
- `drawing_chrome_wiring.rs` — the inspector, the manager and the text
  notes, with `drawing_env` and `DrawingRead`, whose only callers they are.
- `replay_and_history.rs` — the per-frame appliers: replay transport,
  history back-fill, alarms, and the scripted pointer helpers.
- `frame.rs` — `draw_frame`, with `indicator_preview_area`.

What is left is the window's definition and nothing else: the struct, `new`
and `new_with_workspace`, the pickers and `persist_*`, `adopt_tab`,
`arm_strategy_instance`, `attach_surface`, the `eframe::App` impl and
`declare_hooks!` — 1,932 production lines, down 3,456.

No body changed and no method was renamed. The only edits are the
`pub(super)` a child module needs to stay callable from its parent, the
import blocks the seven new files carry, and `app.rs`'s own import list:
names its production code no longer reads but `app::tests` still reaches
through `use super::*` are kept as `#[cfg(test)]` imports, so no file under
`app/tests/` changes. `declare_hooks!` does not move, and the generated
registry and capability inventory are byte-identical.
