+++ crates/app/src/app.rs
@@ -12,59 +12,95 @@
 //! [`ChartPane`], not to the window. The market feeding that pane — channels,
 //! connection state, notices, history counters — is still owned here.
 
-use std::time::{Duration, Instant};
+use std::time::Instant;
 
 use eframe::egui;
 
-use crate::canvas_layout::{MAX_CANVAS_PANES, PaneIdAllocator};
+use crate::canvas_layout::PaneIdAllocator;
 
 mod chart_layers_wiring;
+mod control_host;
 mod demo_hooks;
+mod drawing_chrome_wiring;
 mod drawing_input;
+mod frame;
 mod health;
 mod indicator_manager;
 mod layout_wiring;
+mod menu_bar;
+mod replay_and_history;
+mod tabs;
+mod toolbar_wiring;
 mod workspace_restore;
 mod workspace_save;
-use crate::chart_layers;
-// `ChartLayer` left this file's production code with `maintain_chart_layers`
-// and lives in `app::chart_layers_wiring` now. The one name still read here is
-// read by `heatmap_lamp_on`, which is itself test-only, so the import is gated
-// the same way rather than kept alive by a `chart_layers::` prefix that would
-// appear on exactly one line of the file.
+
+// The tab lifecycle took `saved_context_intervals` with it; `workspace_restore`
+// and `workspace_save` still reach it through `super::`.
+use tabs::saved_context_intervals;
+// Named by the paper-trading and drawing tests through `use super::*`, and by
+// nothing in production outside the module that owns them, so the imports are
+// gated the way the block at the end of this list is.
 #[cfg(test)]
-use crate::chart_layers::ChartLayer;
+use menu_bar::{
+    PAPER_BUY_SHORTCUT, PAPER_CANCEL_SHORTCUT, PAPER_FLATTEN_SHORTCUT, PAPER_REVERSE_SHORTCUT,
+    PAPER_SELL_SHORTCUT,
+};
+#[cfg(test)]
+use replay_and_history::DUPLICATE_OFFSET_BARS;
+
+use crate::chart_layers;
 use crate::config::AppConfig;
-use crate::dock::{Dock, DockEnv, DockTab};
-use crate::drawings::{self, DeleteOutcome, DrawingAuthor};
+use crate::dock::{Dock, DockTab};
+use crate::drawings;
 use crate::feed_notice;
-use crate::harness::{ContextMenuPane, Harness, ScriptedMenu};
+use crate::harness::{Harness, ScriptedMenu};
 use crate::indicator_panel::SettingsDialog;
 use crate::indicator_worker::{IndicatorSource, SlotId};
 use crate::indicators::library::ScriptLibrary;
 use crate::indicators::preset_file;
 use crate::indicators::state_file::{self, SavedKind};
-use crate::loading::{self, LoadingScope, LoadingTask};
-use crate::metrics::{self, FrameStats};
-use crate::pane::{self, ChartPane, DRAWING_ANCHOR_RADIUS_PX, PaneSide};
-use crate::replay_view::{ReplayAction, ReplayView};
+use crate::metrics::FrameStats;
+use crate::pane::{self, PaneSide};
+use crate::replay_view::ReplayView;
 use crate::state::BarSpec;
-use crate::statusbar;
 use crate::style::ChartStyle;
 use crate::symbols_file::{self, AddedSymbols};
-use crate::tab::{CanvasChrome, CanvasLayout, Tab};
-use crate::tabstrip::{self, TabAction};
-use crate::theme;
+use crate::tab::{CanvasLayout, Tab};
 use crate::timezone::TzOffset;
-use crate::toolbar::{self, ToolbarAction};
 use crate::toolrail::{Tool, ToolRail, ToolboxDock};
 use crate::ui_state;
 use crate::window_scale;
 use crate::workspace_store::{LayoutStore, StorePaths, WorkspaceStore};
+use quantick_feed::FeedHandle;
 use quantick_feed::history_reach;
-use quantick_feed::{self as feed, FeedCommand, FeedHandle, ReplayControl};
 use quantick_orderflow::LaneWindow;
-use smallvec::SmallVec;
+
+// Names the window's own code no longer reads: the seven modules above took the
+// last production read of each with them. `app::tests` still reaches every one
+// through `use super::*`, so they stay here as gated imports rather than
+// becoming an edit to `app/tests/` — the same treatment `ChartLayer` already
+// had when `maintain_chart_layers` left for `app::chart_layers_wiring`, and
+// `heatmap_lamp_on` has now taken the last read of that one to `app::tabs`.
+#[cfg(test)]
+use crate::chart_layers::ChartLayer;
+#[cfg(test)]
+use crate::harness::ContextMenuPane;
+#[cfg(test)]
+use crate::loading::LoadingTask;
+#[cfg(test)]
+use crate::metrics;
+#[cfg(test)]
+use crate::pane::{ChartPane, DRAWING_ANCHOR_RADIUS_PX};
+#[cfg(test)]
+use crate::statusbar;
+#[cfg(test)]
+use crate::tabstrip::TabAction;
+#[cfg(test)]
+use crate::theme;
+#[cfg(test)]
+use crate::toolbar::ToolbarAction;
+#[cfg(test)]
+use quantick_feed::{self as feed, FeedCommand, ReplayControl};
 
 /// Id of the tab the window opens with.
 const FIRST_TAB_ID: u64 = 0;
@@ -79,94 +115,6 @@ const DEMO_VISIBLE_SLOTS: usize = 90;
 /// deterministic pair — a native added later must not silently change what
 /// every existing capture shows.
 const AUTOSTART_NATIVES: &[&str] = &["native.ema", "native.cvd"];
-/// Horizontal offset of a duplicated drawing, so the copy is visibly a copy.
-const DUPLICATE_OFFSET_BARS: f32 = 2.0;
-
-/// The slice the drawing chrome reads, assembled from the pieces of the
-/// application it is allowed to see.
-///
-/// A free function rather than a method for the reason
-/// [`indicator_preview_area`] is one: every caller has already split
-/// `QuantickApp` into disjoint borrows to draw a surface through `&mut`, and a
-/// method would want the whole of `self` back. That the compiler insists on
-/// the split is the port working.
-///
-/// `manager_rows` is handed in rather than gathered here. Only one of the two
-/// call sites draws the list, and building a row per object for the site that
-/// does not would be a per-frame allocation for a window nobody is looking at.
-fn drawing_env<'a>(
-    tab: &'a Tab,
-    toolrail: &ToolRail,
-    presets: &'a drawings::presets::PresetStore,
-    read: DrawingRead<'a>,
-) -> crate::surfaces::DrawingEnv<'a> {
-    let side = tab.drawing_side();
-    let pane = tab.pane(side);
-    let selected = pane.drawings.selected().and_then(|index| {
-        pane.drawings
-            .items()
-            .get(index)
-            .map(|drawing| crate::surfaces::drawing_chrome::SelectedDrawing { index, drawing })
-    });
-    crate::surfaces::DrawingEnv {
-        selected,
-        chart_area: pane.last_chart_area,
-        focused_chart_area: tab.focused_pane().last_chart_area,
-        lane_divider_x: pane.last_lane_divider_x,
-        auto_range: pane.last_auto_range,
-        selected_bbox: read.selected_bbox,
-        selected_band: read.selected_band,
-        tab: tab.id,
-        side,
-        drawing_tool_armed: matches!(toolrail.tool(), Tool::Drawing(_)),
-        toolbox_dock: toolrail.dock(),
-        authored_objects: read.authored_objects,
-        manager_rows: read.manager_rows,
-        presets,
-    }
-}
-
-/// The parts of [`drawing_env`] that cost something to work out, gathered by
-/// the caller so each pass pays only for what it draws.
-///
-/// Three fields and three prices. Projecting the selection's painted bounds
-/// walks its anchors through the price scale; naming its band formats a
-/// string; counting an assistant's objects walks every pane of every tab. All
-/// three are per-frame while a selection is on screen, which is why the pass
-/// that only runs the capture hooks gathers none of them and says so.
-#[derive(Default)]
-struct DrawingRead<'a> {
-    selected_bbox: Option<egui::Rect>,
-    selected_band: Option<String>,
-    authored_objects: usize,
-    manager_rows: &'a [crate::surfaces::drawing_chrome::ManagerRow],
-}
-
-/// The chart rectangle a settings dialog is previewing an unapplied draft on,
-/// if one is.
-///
-/// The pane the *dialog* was opened over, not the focused one: a trader can
-/// preview a curve on the left pane and then click the right, and the banner
-/// belongs over the numbers that are actually provisional.
-///
-/// A free function rather than a method because its only caller has already
-/// split `QuantickApp` into disjoint borrows to build the surface
-/// environment, and a method would want the whole of `self` back. Per-frame,
-/// and shaped to leave immediately: no dialog open — the ordinary case — is
-/// one `Option` test before the tab scan is reached.
-fn indicator_preview_area(
-    tabs: &[Tab],
-    dialog: Option<&SettingsDialog>,
-    target: TabSlot,
-) -> Option<egui::Rect> {
-    if !dialog.is_some_and(|dialog| dialog.previewed) {
-        return None;
-    }
-    tabs.iter()
-        .find(|tab| tab.id == target.tab)
-        .map(|tab| tab.pane(target.side))
-        .and_then(|pane| pane.last_chart_area)
-}
 
 /// Format the forming bar's countdown, e.g. `37/50 ticks`.
 ///
@@ -1621,519 +1569,6 @@ impl QuantickApp {
         crate::paper_state::save(&path, &state);
     }
 
-    /// The active tab beside the config it reads.
-    ///
-    /// Split here, once, because almost every tab operation needs both and
-    /// `self.tabs[i].f(&self.config)` is a borrow error at every call site.
-    fn active_with_config(&mut self) -> (&mut Tab, &AppConfig) {
-        (&mut self.tabs[self.active_tab], &self.config)
-    }
-
-    /// The tab on screen.
-    fn active_tab(&self) -> &Tab {
-        &self.tabs[self.active_tab]
-    }
-
-    /// See [`Self::active_tab`].
-    fn active_tab_mut(&mut self) -> &mut Tab {
-        &mut self.tabs[self.active_tab]
-    }
-
-    /// Read-only application roots available to the on-demand control
-    /// projections. The gateway never receives `QuantickApp`; it receives the
-    /// owned DTOs built from these narrow views.
-    /// One tab by position, for a control capability that resolved an id.
-    pub(crate) fn control_tab_at(&self, index: usize) -> Option<&Tab> {
-        self.tabs.get(index)
-    }
-
-    /// The mutable twin, for the cockpit tier.
-    ///
-    /// Narrow on purpose: the layout capabilities need to *change* a tab, and
-    /// handing them the whole application would let a later one reach past the
-    /// canvas into the feed or the simulator.
-    pub(crate) fn control_tab_at_mut(&mut self, index: usize) -> Option<&mut Tab> {
-        self.tabs.get_mut(index)
-    }
-
-    /// One tab beside the configuration it reads, by position.
-    ///
-    /// [`Self::active_with_config`] for a tab that is not necessarily the
-    /// active one — a capability names the tab it acts on, and respawning a
-    /// feed needs the feed table the same way a click in the corner does.
-    pub(crate) fn control_tab_with_config(
-        &mut self,
-        index: usize,
-    ) -> Option<(&mut Tab, &AppConfig)> {
-        let Self { tabs, config, .. } = self;
-        tabs.get_mut(index).map(|tab| (tab, &*config))
-    }
-
-    /// The trading host of the tab on screen — where the `trade.*` actions
-    /// land. The active tab and not an addressed one: an order belongs to
-    /// the symbol the trader is looking at, and a call that could quietly
-    /// trade a chart nobody has open is a call nobody should be able to
-    /// make.
-    pub(crate) fn control_active_paper_mut(
-        &mut self,
-    ) -> Option<&mut crate::paper_trading::PaperTrading> {
-        // Fallible, because the rest of the control code does not trust the
-        // invariant either: `annotate::resolve_target` guards an empty tab
-        // list and clamps the index, and two more sites clamp it. A
-        // `trade.*` call must answer "this window has no chart open" rather
-        // than panic the whole trading application, and it must resolve the
-        // *same* tab its own read-back resolves.
-        self.tabs.get_mut(self.active_tab).map(|tab| &mut tab.paper)
-    }
-
-    /// The read side of [`Self::control_active_paper_mut`], resolved the same
-    /// way so a call and its read-back can never name different tabs.
-    pub(crate) fn control_active_paper(&self) -> Option<&crate::paper_trading::PaperTrading> {
-        self.tabs.get(self.active_tab).map(|tab| &tab.paper)
-    }
-
-    pub(crate) fn control_tabs(&self) -> &[Tab] {
-        &self.tabs
-    }
-
-    pub(crate) fn control_active_tab_index(&self) -> usize {
-        self.active_tab
-    }
-
-    /// Open the assistant's popup. One at a time: a second message replaces
-    /// the first rather than stacking windows over a chart someone is
-    /// trading, and the trader dismisses it.
-    pub(crate) fn show_agent_popup(&mut self, popup: crate::control::AgentPopup) {
-        self.surfaces.agent_popup.show(popup);
-    }
-
-    /// Post one line to the window's own acknowledgement lane — the same
-    /// channel a delete or a workspace save uses, with no Undo: there is
-    /// nothing to take back from having been told something.
-    pub(crate) fn show_agent_toast(&mut self, message: String) {
-        self.surfaces.toast.note(message, Instant::now());
-    }
-
-    /// Ask for the platform's attention sound, through the same sink the
-    /// alarms use, and report honestly when it could not be made rather
-    /// than letting a client believe it was heard.
-    pub(crate) fn sound_agent_alert(&mut self) -> Option<String> {
-        self.alerts
-            .play(&[crate::audio::Cue::default()])
-            .err()
-            .map(ToOwned::to_owned)
-    }
-
-    /// One pane, by tab position and side — the mutable half of
-    /// [`Self::control_tabs`], for the actions that place objects.
-    pub(crate) fn control_pane_mut(
-        &mut self,
-        tab_index: usize,
-        side: crate::pane::PaneSide,
-    ) -> &mut ChartPane {
-        self.tabs[tab_index].pane_mut(side)
-    }
-
-    /// What a freshly placed object of `tool` opens with, through the same
-    /// door the click path uses — saved defaults, named preset and all.
-    pub(crate) fn control_new_drawing(&self, tool: drawings::DrawingTool) -> drawings::NewDrawing {
-        drawings::new_drawing_from_defaults(&self.drawing_presets, tool)
-    }
-
-    pub(crate) fn control_config(&self) -> &AppConfig {
-        &self.config
-    }
-
-    /// The window's footprint setup — the one a pane falls back to when it
-    /// carries no override of its own.
-    pub(crate) fn control_footprint_config(&self) -> &crate::footprint_config::FootprintConfig {
-        &self.footprint_config
-    }
-
-    /// The window's shared chart style, which owns the layers no pane does.
-    pub(crate) fn control_style(&self) -> &ChartStyle {
-        &self.style
-    }
-
-    /// The drawing tool rail: which tool is armed, and whether it is on
-    /// screen at all.
-    pub(crate) fn control_tool_rail(&self) -> &ToolRail {
-        &self.toolrail
-    }
-
-    /// The colour the chart's corner is wearing, or `None` while the chart
-    /// is being fed.
-    ///
-    /// The status line's provenance dot takes this rather than deciding for
-    /// itself. It used to read the connection alone, which is a socket's
-    /// opinion: a terminal that froze with the socket open had the
-    /// bottom-left of the window saying `live` while the bottom-right said
-    /// `offline`, about the same feed, at the same moment. Two surfaces
-    /// disagreeing about the one question the trader is asking is worse than
-    /// either answer alone, so there is one report and both read it.
-    fn feed_offline_accent(
-        &self,
-        stall: Option<&quantick_feed::stall::Stall>,
-    ) -> Option<egui::Color32> {
-        feed_notice::report(&self.active_tab().notice, stall)
-            .filter(feed_notice::Report::is_offline)
-            .map(|report| report.accent())
-    }
-
-    /// Where the feed's offline chip was painted, or `None` when it was not.
-    ///
-    /// The projection reads what was drawn rather than re-deciding it, so the
-    /// scene and the screen cannot disagree across the edge of a stall budget.
-    pub(crate) fn control_feed_chip_rect(&self) -> Option<egui::Rect> {
-        self.feed_chip_rect
-    }
-
-    /// Whether the recovery popup that chip opens is showing, on the chart
-    /// the trader is looking at.
-    pub(crate) fn control_feed_popup_open(&self) -> bool {
-        self.feed_popup_tab == Some(self.active_tab().id)
-    }
-
-    /// The right-hand dock: whether it is shown, and which tab is open.
-    pub(crate) fn control_dock(&self) -> &Dock {
-        &self.dock
-    }
-
-    pub(crate) fn control_timezone(&self) -> TzOffset {
-        self.tz
-    }
-
-    pub(crate) fn control_workspace_flags(&self) -> (bool, bool, bool) {
-        (
-            self.workspace.session().save_on_exit(),
-            self.show_perf,
-            self.progressive_history,
-        )
-    }
-
-    /// Choose how far one press of *load older* reaches.
-    ///
-    /// The named call behind the history menu's reach chips and the
-    /// `QUANTICK_HISTORY_REACH` hook — one path, so an operator without a
-    /// mouse sets what a click sets. Mirrored onto every tab by `drain_tabs`,
-    /// where a run in flight also reads it: withdrawing the longer reach is
-    /// how a trader calls that run off.
-    pub(crate) fn set_history_reach(&mut self, reach: history_reach::HistoryReach) {
-        self.history_reach = reach;
-    }
-
-    /// How far back one press of the `by time` reach pulls, in minutes of
-    /// traded time.
-    ///
-    /// Clamped rather than refused: a span of zero is a press that asks for
-    /// nothing, and the operator that sent it meant *some* history. The
-    /// ceiling is the campaign's own span cap, past which no run can reach
-    /// anyway, so accepting a larger number would be promising a reach the
-    /// budgets forbid.
-    pub(crate) fn set_history_reach_span_minutes(&mut self, minutes: u32) {
-        let ceiling = (history_reach::MAX_CAMPAIGN_SPAN_MS / 60_000) as u32;
-        self.history_reach_span_minutes = minutes.clamp(1, ceiling);
-    }
-
-    /// What that span is now, for an operator reading back what it set.
-    pub(crate) fn control_history_reach_span_minutes(&self) -> u32 {
-        self.history_reach_span_minutes
-    }
-
-    /// How far the window's *load older* press reaches, and whether a chart
-    /// cut by trades carries the venue's candles.
-    ///
-    /// Both are choices an operator without a mouse has to be able to read
-    /// back after setting them — the reach especially, since it decides
-    /// whether one press is one request or a run of them.
-    pub(crate) fn control_history_settings(&self) -> (history_reach::HistoryReach, bool) {
-        (self.history_reach, self.venue_lead_in)
-    }
-
-    /// Whether a recording opens with the session day before it joined in
-    /// front, and a download fetches that day's tape too.
-    ///
-    /// A choice an operator without a mouse has to be able to read back after
-    /// setting it: it decides what a replay they are about to open will hold.
-    pub(crate) fn control_replay_day_before(&self) -> bool {
-        self.replay_view.day_before()
-    }
-
-    /// Invoke one registered control action from inside the application,
-    /// attributed to the human at this window (or to automation when a
-    /// control trace replays it). The hotkey, the `QUANTICK_CONTROL_MARK`
-    /// hook and the tests all arrive here; there is no second path.
-    pub(crate) fn control_action(
-        &mut self,
-        capability_id: &str,
-        capability_version: u32,
-        origin: crate::control::ActionOrigin,
-        input: serde_json::Value,
-    ) -> Result<serde_json::Value, quantick_control::error::ControlError> {
-        let Some(mut access) = self.control_access.take() else {
-            return Err(quantick_control::error::ControlError::invalid_request(
-                "control access is not installed",
-            ));
-        };
-        let outcome =
-            access.invoke_local_action(self, capability_id, capability_version, input, origin);
-        self.control_access = Some(access);
-        outcome
-    }
-
-    /// How many objects an operator other than the trader placed, across
-    /// every pane one can reach — an assistant may annotate any open tab, so
-    /// counting the active pane alone would offer to take back a subset and
-    /// call it all of them.
-    fn authored_object_count(tabs: &[Tab]) -> usize {
-        tabs.iter()
-            .map(|tab| {
-                tab.panes()
-                    .map(|(pane, _side)| pane.drawings.authored_count())
-                    .sum::<usize>()
-            })
-            .sum()
-    }
-
-    /// Take back every object an operator placed, wherever it is. One undo
-    /// entry per pane, and the resting orders of any armed strategy go with
-    /// the objects they were anchored to.
-    fn remove_every_authored_object(&mut self) -> usize {
-        let mut removed = 0;
-        for tab in &mut self.tabs {
-            // Every pane the tab holds, not the two it used to. "Remove
-            // objects placed for you" promises to take them *all* back, and a
-            // sweep that skipped the second stacked chart would leave an
-            // assistant's marks behind while reporting the job done.
-            for pane in tab.panes_mut() {
-                let taken = pane.drawings.remove_authored();
-                if taken > 0 {
-                    pane.sweep_strategy_orphans();
-                    removed += taken;
-                }
-            }
-        }
-        removed
-    }
-
-    /// The annotate tier's launch hooks: one agent-authored label, one
-    /// notification. Both go through the registered action with an agent
-    /// actor — the same path the gateway takes for a remote client — so what
-    /// a screenshot shows is what a real assistant would have produced.
-    fn apply_control_annotate_hooks(&mut self) {
-        if let Some(text) = self.pending_control_annotation.take() {
-            let anchor = {
-                let pane = self.active_tab().drawing_pane();
-                let slot = pane.slots().saturating_sub(1);
-                match (pane.slot_open_time(slot), pane.closed_bar(slot)) {
-                    (Some(time), Some(bar)) => Some(serde_json::json!({
-                        "time_unix_ms": time,
-                        "price": rust_decimal::prelude::ToPrimitive::to_f64(&bar.close)
-                            .unwrap_or(1.0)
-                            .to_string(),
-                    })),
-                    // No bars yet: put the hook back and take it next frame,
-                    // rather than annotating a chart that has nothing on it.
-                    _ => {
-                        self.pending_control_annotation = Some(text.clone());
-                        None
-                    }
-                }
-            };
-            if let Some(anchor) = anchor {
-                self.pending_control_annotation = None;
-                self.run_hook_action(
-                    "annotate.label.create",
-                    serde_json::json!({ "anchors": [anchor], "text": text }),
-                );
-            }
-        }
-        if let Some(request) = self.pending_control_notification.take() {
-            let (channel, message) = request
-                .split_once(':')
-                .unwrap_or(("toast", request.as_str()));
-            let capability = match channel.trim() {
-                "popup" => Some("notify.popup"),
-                "sound" => Some("notify.sound"),
-                "toast" => Some("notify.toast"),
-                other => {
-                    tracing::warn!(
-                        target: "quantick::control",
-                        event_code = "CONTROL_NOTIFY_HOOK_REFUSED",
-                        channel = other,
-                        "QUANTICK_CONTROL_NOTIFY names no notification channel"
-                    );
-                    None
-                }
-            };
-            if let Some(capability) = capability {
-                self.run_hook_action(
-                    capability,
-                    serde_json::json!({ "message": message, "title": "From your assistant" }),
-                );
-            }
-        }
-    }
-
-    /// Invoke one registered action as an *agent* would, from inside this
-    /// window. The hooks use it so a screenshot shows a real assistant's
-    /// object, attribution and all, without a client on the socket.
-    fn run_agent_action(
-        &mut self,
-        capability_id: &str,
-        input: serde_json::Value,
-    ) -> Result<serde_json::Value, quantick_control::error::ControlError> {
-        let Some(mut access) = self.control_access.take() else {
-            return Err(quantick_control::error::ControlError::invalid_request(
-                "control access is not installed",
-            ));
-        };
-        // No identity, no actor to sign with: the same structured refusal an
-        // action gets, rather than a panic on the first frame.
-        let Some(actor) = access.hook_agent_actor() else {
-            self.control_access = Some(access);
-            return Err(quantick_control::error::ControlError::invalid_request(
-                "this window has no control identity to act with",
-            ));
-        };
-        let outcome = access.invoke_local_action(
-            self,
-            capability_id,
-            1,
-            input,
-            crate::control::ActionOrigin::Remote(Box::new(actor)),
-        );
-        self.control_access = Some(access);
-        outcome
-    }
-
-    /// A launch hook's action, with its failure reported where a scripted run
-    /// will see it: the hook is fire-and-forget, so nothing else would.
-    fn run_hook_action(&mut self, capability_id: &str, input: serde_json::Value) {
-        if let Err(error) = self.run_agent_action(capability_id, input) {
-            tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_HOOK_ACTION_FAILED",
-                capability = capability_id,
-                error_code = %error.code,
-                error = %error.message,
-                "an annotate hook could not run its action"
-            );
-        }
-    }
-
-    /// The mark hotkey's body: `attention.mark.create` with the resolved
-    /// cursor target, attributed to the human.
-    pub(crate) fn take_mark(&mut self, note: Option<String>) {
-        let mut input = serde_json::Map::new();
-        if let Some(note) = note {
-            input.insert("note".to_owned(), serde_json::Value::String(note));
-        }
-        // No target: the action port resolves the pointer at the moment of
-        // the gesture and records the resolved input, so the trace line
-        // determines the mark on its own and a rerun marks the same bar.
-        match self.control_action(
-            crate::control::MARK_CAPABILITY_ID,
-            crate::control::MARK_CAPABILITY_VERSION,
-            crate::control::ActionOrigin::Human,
-            serde_json::Value::Object(input),
-        ) {
-            Ok(result) => tracing::info!(
-                target: "quantick::control",
-                event_code = "CONTROL_MARK_TAKEN",
-                sequence = %result["sequence"],
-                "mark taken"
-            ),
-            Err(error) => tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_MARK_REFUSED",
-                code = %error.code,
-                "mark refused"
-            ),
-        }
-    }
-
-    pub(crate) fn control_frame_metrics(&self) -> ControlFrameMetrics {
-        ControlFrameMetrics {
-            wall_average_ms: self.frames.avg_ms(),
-            wall_worst_ms: self.frames.worst_ms(),
-            frames_per_second: self.frames.fps(),
-            cpu_average_ms: self.cpu_frames.avg_ms(),
-            cpu_worst_ms: self.cpu_frames.worst_ms(),
-        }
-    }
-
-    /// The pane the chrome speaks for: the active tab's focused pane (§11).
-    fn focused_pane(&self) -> &ChartPane {
-        self.active_tab().focused_pane()
-    }
-
-    /// See [`Self::focused_pane`].
-    fn focused_pane_mut(&mut self) -> &mut ChartPane {
-        self.active_tab_mut().focused_pane_mut()
-    }
-
-    /// The pane every drawing surface speaks for: the one holding the
-    /// selection, which is the focused pane unless a shared mark was taken
-    /// from the chart it is mirrored on (see [`Tab::drawing_side`]).
-    ///
-    /// The inspector, the keyboard, the object manager and the toast all read
-    /// through here, so an object selected on either of its two charts is
-    /// edited and deleted from either of them.
-    fn drawing_pane(&self) -> &ChartPane {
-        self.active_tab().drawing_pane()
-    }
-
-    /// See [`Self::drawing_pane`].
-    fn drawing_pane_mut(&mut self) -> &mut ChartPane {
-        self.active_tab_mut().drawing_pane_mut()
-    }
-
-    /// The slot a command from the chrome addresses: the active tab, its
-    /// focused pane, that slot.
-    fn target_slot(&self, slot: SlotId) -> TabSlot {
-        TabSlot {
-            tab: self.active_tab().id,
-            side: self.active_tab().focused_side(),
-            slot,
-        }
-    }
-
-    /// Open `feed_id`/`symbol` in a new tab and make it active.
-    ///
-    /// Opening a market a tab already holds is allowed — two views of one
-    /// book are a legitimate thing to want. For MetaTrader that means two
-    /// listeners on one port, and the second one loses the bind: that tab
-    /// shows the bridge's own bind-failure notice, which is the honest answer
-    /// and the reason `[metatrader.ports]` maps a port per symbol.
-    fn open_tab(&mut self, feed_id: String, symbol: String, spec: Option<BarSpec>) {
-        let Some(provider) = self.config.provider_of(&feed_id) else {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "TAB_OPEN_UNKNOWN_FEED",
-                feed = %feed_id,
-                action = "ignore_request",
-                "asked to open a feed the config does not have"
-            );
-            return;
-        };
-        // One feed per tab, resolved per symbol: a MetaTrader tab binds the
-        // port `[metatrader.ports]` maps its symbol to (`endpoint_for`), so two
-        // MT5 tabs on different symbols listen on different ports and each
-        // finds its own bridge. Two tabs on the *same* MT5 symbol is allowed
-        // and means one port for two listeners: the second loses the bind and
-        // shows the feed's own MT5_BIND_FAILED notice, which is the honest
-        // answer rather than a silently dead chart.
-        let handle = feed::spawn_live(
-            provider,
-            &symbol,
-            &self.config.metatrader,
-            crate::paper_home::shelf_dir(),
-        );
-        self.adopt_tab(feed_id, symbol, handle, spec);
-    }
-
     /// Take a market that is already streaming as a new tab, and make it the
     /// active one.
     ///
@@ -2258,377 +1693,6 @@ impl QuantickApp {
         }
     }
 
-    /// Close the tab at `index`, activating a neighbour.
-    ///
-    /// The last tab stays: a window with no market has nothing to draw. What
-    /// the closed tab owned goes with it — dropping its `FeedHandle` closes
-    /// the receivers its feed thread sends into, and dropping its panes drops
-    /// the indicator worker and book worker handles, whose run loops end when
-    /// their command channels disconnect. No joins, no shutdown protocol.
-    fn close_tab(&mut self, index: usize) {
-        if self.tabs.len() <= 1 || index >= self.tabs.len() {
-            return;
-        }
-        let mut closed = self.tabs.remove(index);
-        // The tab's session ends here. Everything else it owns can simply be
-        // dropped — the feed thread and the workers stop when their channels
-        // go — but a simulated position is state the user created, and the
-        // paper-trading contract says it ends in a labeled, journaled flatten,
-        // never by vanishing with its window.
-        closed.close();
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "TAB_CLOSED",
-            tab = closed.id,
-            feed = %closed.feed_id,
-            symbol = %closed.symbol,
-            tabs = self.tabs.len(),
-            action = "drop_feed_and_workers",
-            "closing a market tab"
-        );
-        // Its slots are gone with its panes; the bookkeeping must not outlive
-        // them or a later tab reusing a slot number would inherit its kind.
-        self.slot_kinds.retain(|(owner, _)| owner.tab != closed.id);
-        self.operator_slots.retain(|owner| owner.tab != closed.id);
-        self.script_files
-            .retain(|(owner, ..)| owner.tab != closed.id);
-        self.pending_hidden.retain(|owner| owner.tab != closed.id);
-        self.pending_styles
-            .retain(|(owner, _)| owner.tab != closed.id);
-        self.active_tab = self.active_tab.min(self.tabs.len() - 1);
-        drop(closed);
-    }
-
-    /// Move `delta` tabs along the strip, wrapping (§10: Ctrl+Tab).
-    fn cycle_tab(&mut self, delta: isize) {
-        if self.tabs.len() < 2 {
-            return;
-        }
-        let count = self.tabs.len() as isize;
-        let next = (self.active_tab as isize + delta).rem_euclid(count);
-        self.active_tab = next as usize;
-    }
-
-    /// Whether the toolbar's heatmap lamp is lit.
-    ///
-    /// The *switch*, not what capture lets through it — the same reading the
-    /// layer file was taught in 848cba0, and for a sibling reason. A lamp lit
-    /// from `depth_visible()` (`enabled && show_depth`) reports the heatmap off
-    /// for as long as book capture is starting, and forever on a source with no
-    /// book: the trader sees an unlit button, presses it, and switches the
-    /// layer they wanted *off*. The button already has an honest way to say a
-    /// source cannot fill it — `.enabled(...)` carrying its
-    /// `disabled_explanation` — so the lamp beside it answers the only other
-    /// question there is.
-    ///
-    /// A named reading rather than an expression inside the toolbar's own
-    /// frame, so the rule can be asserted without painting a toolbar. What
-    /// reads it back without looking at the screen is the semantic scene,
-    /// which takes the same `Tab::layer_toggle_state` this delegates to.
-    #[must_use]
-    #[cfg(test)]
-    fn heatmap_lamp_on(&self) -> bool {
-        // Through the group's one reading, so this named rule and the lamp the
-        // toolbar actually paints cannot become two answers to one question.
-        // `#[cfg(test)]` because the toolbar now takes the group's reading
-        // directly: keeping a second production entry point to the same answer
-        // is how the two drift.
-        self.active_tab()
-            .layer_toggle_state(
-                ChartLayer::Heatmap,
-                &self.style,
-                self.active_tab().capabilities(&self.config),
-            )
-            .0
-    }
-
-    /// Build the toolbar's model from the app's state, draw it, and carry
-    /// out whatever it asked (§6 — the toolbar module owns grouping and the
-    /// overflow rule; this method owns the side effects).
-    fn draw_toolbar(&mut self, ctx: &egui::Context) {
-        // Pre-collect owned option lists so the toolbar's combos don't borrow
-        // `self.config` while they mutate `self.feed_id` / `self.active_tab().symbol`.
-        // Providers that aren't streaming yet are labelled "(soon)" so the
-        // menu is honest about what actually connects.
-        let feeds: Vec<(String, String)> = self
-            .config
-            .feeds
-            .iter()
-            .map(|f| {
-                let label = if f.provider.is_implemented() {
-                    f.name.clone()
-                } else {
-                    format!("{} (soon)", f.name)
-                };
-                (f.id.clone(), label)
-            })
-            .collect();
-        let symbols: Vec<String> = self
-            .config
-            .feed(&self.active_tab().feed_id)
-            .map(|f| f.symbols.clone())
-            .unwrap_or_default();
-        // During a replay the SOURCE group gives way to what is actually
-        // playing: a live venue cannot be picked without leaving the
-        // recording first, and a combo that silently did so would throw away
-        // the session mid-run.
-        let replay = self
-            .active_tab()
-            .replay
-            .as_ref()
-            .map(|link| toolbar::ReplaySource {
-                label: link.label(),
-                hover: format!(
-                    "Replaying {}\nSide source: {}",
-                    link.session.path.display(),
-                    link.session
-                        .header
-                        .side_source
-                        .as_deref()
-                        .unwrap_or("not recorded"),
-                ),
-            });
-        let capabilities = self.active_tab().capabilities(&self.config);
-        let candles_held = self.active_tab().venue_candles_held();
-        let older_candles = self.active_tab().older_candles(capabilities);
-        let feed_display_name = self.active_tab().feed_display_name(&self.config).to_owned();
-        // One reading per lamp, taken through the call the semantic scene
-        // makes too, so the button and what an operator captures cannot
-        // disagree about a layer. Every lamp reports the *switch* rather than
-        // what the source lets through it — the rule `heatmap_lamp_on` names,
-        // now the whole group's.
-        let layers = toolbar::LayerToggle::ALL.map(|toggle| {
-            let (on, blocked) =
-                self.active_tab()
-                    .layer_toggle_state(toggle.layer(), &self.style, capabilities);
-            toolbar::LayerToggleState { on, blocked }
-        });
-        // The focused pane's slots (§11): the menu lists what a command from
-        // it would act on, and never the pane beside it.
-        let indicators: Vec<toolbar::IndicatorMenuEntry> = self
-            .focused_pane()
-            .indicators
-            .all()
-            .iter()
-            .map(|view| toolbar::IndicatorMenuEntry {
-                slot: view.slot.0,
-                label: view.label().to_owned(),
-                hidden: view.hidden,
-                errored: view.error.is_some(),
-                stale: view.stale.is_some(),
-            })
-            .collect();
-        let scripts: Vec<String> = self
-            .script_library
-            .entries()
-            .iter()
-            .map(|entry| entry.name.clone())
-            .collect();
-        let dock_visible = self.dock.visible();
-        let show_style = self.surfaces.style_panel.is_open();
-        // Read before the tab is borrowed mutably. The reach is the window's
-        // standing choice, like the progressive-history switch — a trader who
-        // picked "previous session" once means it in the next tab too — so it
-        // is split off and written back the way the layout picker's flags are.
-        let history_reach_running = self.active_tab().history_reach_running();
-        let mut history_reach = self.history_reach;
-        let mut history_reach_span_minutes = self.history_reach_span_minutes;
-        let mut history_menu_rect = self.history_menu_rect;
-        // The SOURCE group writes straight into the active tab: a feed or
-        // symbol change is that tab's market switch. The BARS group writes
-        // into the *focused pane* — the pane the status bar reads and every
-        // indicator command lands on (§11) — so the three chrome surfaces
-        // can never disagree about which chart a command describes, and in
-        // the Time layout the group governs the chart actually on screen.
-        // Split off the picker's flags before the tab borrow: the model wants
-        // both, and they live on the same struct.
-        let mut layout_picker_open = self.layout_picker_open;
-        // One shot: the hook opens the popover on the first drawn frame and
-        // then gets out of the way, so a trader's click can close it.
-        let layout_picker_autostart = self.harness.take_layout_picker_autostart();
-        let tab = self.active_tab_mut();
-        let focused = tab.focused_side();
-        let pane = match focused {
-            PaneSide::Time(slot) => tab.time_panes.get_mut(slot).unwrap_or(&mut tab.flow_pane),
-            PaneSide::Flow => &mut tab.flow_pane,
-        };
-        let mut model = toolbar::ToolbarModel {
-            layout_preset: Some(tab.layout.preset()),
-            layout_picker_open: &mut layout_picker_open,
-            layout_picker_request_open: layout_picker_autostart,
-            feeds,
-            feed_id: &mut tab.feed_id,
-            feed_display_name,
-            symbols,
-            symbol: &mut tab.symbol,
-            replay,
-            kind: &mut pane.kind,
-            tick_n: &mut pane.tick_n,
-            volume_units: &mut pane.volume_units,
-            dollar_notional: &mut pane.dollar_notional,
-            time_interval_ms: &mut pane.time_interval_ms,
-            imbalance_target: &mut pane.imbalance_target,
-            imbalance_unit: &mut pane.imbalance_unit,
-            history_step: &mut tab.history_step,
-            history_menu_rect: &mut history_menu_rect,
-            history_reach_span_minutes: &mut history_reach_span_minutes,
-            history_reach: &mut history_reach,
-            history_reach_running,
-            history_trades: tab.history_trades,
-            history_candles: candles_held,
-            older_candles,
-            capabilities,
-            layers,
-            dock_visible,
-            appearance_open: show_style,
-            paper: toolbar::PaperTradeModel {
-                // The lock reaches the toolbar too. Gating only the dock's
-                // pair left these lit while the ticket refused, so a fast
-                // click here only toasted - and the doc promises the entry
-                // pair disables.
-                ready: tab.paper.ready() && !tab.paper.risk_report().1,
-                buy_label: tab.paper.entry_label(quantick_engine::Side::Buy),
-                sell_label: tab.paper.entry_label(quantick_engine::Side::Sell),
-                buy_hover: tab.paper.entry_hover(quantick_engine::Side::Buy),
-                sell_hover: tab.paper.entry_hover(quantick_engine::Side::Sell),
-                close_label: tab.paper.close_button_label(),
-            },
-            indicators,
-            scripts,
-        };
-        let actions = toolbar::draw(ctx, &mut model);
-        // The popover's own state, back where it lives. Without this the flag
-        // resets every frame and the button never reads as open.
-        drop(model);
-        self.layout_picker_open = layout_picker_open;
-        self.set_history_reach(history_reach);
-        // Through the setter, so a value dragged past the campaign's own span
-        // cap is clamped in the one place that knows the cap.
-        self.set_history_reach_span_minutes(history_reach_span_minutes);
-        self.history_menu_rect = history_menu_rect;
-        // A newly picked feed may not offer the current symbol. Never during
-        // a replay: the recorded instrument belongs to no live feed's menu,
-        // and snapping it away would relabel the whole session — the status
-        // bar and the logs must keep naming what is actually playing.
-        if self.active_tab().replay.is_none() {
-            let (tab, config) = self.active_with_config();
-            tab.ensure_symbol_valid(config);
-            tab.refresh_chip_label(config);
-        }
-        for action in actions {
-            self.apply_toolbar_action(action);
-        }
-    }
-
-    /// Switch the active tab to `preset`.
-    ///
-    /// **The one path.** The toolbar's picker, the `View → Layout` menu and
-    /// the keyboard all arrive here, so none of them can grow its own idea of
-    /// what applying a layout does. A control-plane capability joins them by
-    /// calling this, never by repeating it.
-    fn apply_layout_preset(&mut self, preset: &'static crate::canvas_layout::LayoutPreset) {
-        let Some(layout) = CanvasLayout::from_preset(preset) else {
-            // A preset the canvas cannot draw yet is refused rather than
-            // approximated: switching to the nearest arrangement would be the
-            // picker showing one layout and the canvas drawing another.
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "LAYOUT_PRESET_UNSUPPORTED",
-                preset = %preset.id,
-                action = "layout_left_as_is",
-                "the layout registry names an arrangement the canvas cannot draw yet"
-            );
-            return;
-        };
-        self.active_tab_mut().set_layout(layout);
-    }
-
-    /// One toolbar side effect. Layer toggles reuse the same code paths the
-    /// old checkboxes took, so provider gating and command acknowledgement
-    /// rules are unchanged.
-    fn apply_toolbar_action(&mut self, action: ToolbarAction) {
-        match action {
-            ToolbarAction::LoadOlder => {
-                let (tab, config) = self.active_with_config();
-                tab.request_older_history(config);
-            }
-            ToolbarAction::LoadOlderCandles => {
-                // Read before the tab is borrowed mutably — and the capability
-                // block rather than the whole config, because that is all the
-                // request needs to know.
-                let capabilities = self.active_tab().capabilities(&self.config);
-                self.active_tab_mut()
-                    .request_older_ohlcv_history(capabilities);
-            }
-            ToolbarAction::SetHeatmap(shown) => {
-                self.active_tab_mut().tape_mut().set_depth_visible(shown);
-            }
-            ToolbarAction::SetBubbles(enabled) => {
-                self.active_tab_mut()
-                    .tape_mut()
-                    .set_bubbles_enabled(enabled);
-            }
-            ToolbarAction::SetLiveStrip(shown) => {
-                self.active_tab_mut().flow_pane.live_strip_visible = shown;
-            }
-            // The focused pane's own field, through the same setter the pane's
-            // layer menu calls — so the button, the menu and the lamp can
-            // never disagree about which chart the command described.
-            ToolbarAction::SetFootprint(shown) => {
-                self.focused_pane_mut().footprint_visible = shown;
-            }
-            ToolbarAction::OpenFootprintSettings => self.surfaces.footprint_settings.open(),
-            ToolbarAction::OpenDockTab(tab) => self.dock.open_tab(tab),
-            ToolbarAction::SetLayout(preset) => self.apply_layout_preset(preset),
-            ToolbarAction::ToggleDock => self.dock.toggle_visible(),
-            ToolbarAction::ToggleAppearance => self.surfaces.style_panel.toggle(),
-            // Every indicator command lands on the focused pane (§11), which
-            // is the flow pane whenever the canvas is not split.
-            // Adding an indicator by hand is the plainest possible request to
-            // see one, so it opens a folded legend rather than letting the new
-            // row land inside the puck — the trader would get one more dot and
-            // no way to tell the add from a no-op. Not the auto-collapse the
-            // design ruled out: that rule protects against hiding what nobody
-            // asked to hide, and unfolding hides nothing. It lives on this
-            // path, the trader's own, and not in `ChartPane::add_indicator`,
-            // which the workspace restore and the harness hooks also travel —
-            // there it would erase the fold on every launch.
-            ToolbarAction::AddNative(id) => {
-                self.set_focused_legend_collapsed(false);
-                self.add_native_indicator(id);
-            }
-            ToolbarAction::ToggleIndicatorHidden(slot) => {
-                let target = self.target_slot(SlotId(slot));
-                self.toggle_indicator_hidden_at(target);
-            }
-            ToolbarAction::RemoveIndicator(slot) => {
-                let target = self.target_slot(SlotId(slot));
-                self.remove_indicator_at(target);
-            }
-            ToolbarAction::AddScriptIndicator(index) => {
-                self.set_focused_legend_collapsed(false);
-                self.add_script_indicator(index);
-            }
-            ToolbarAction::OpenIndicatorSettings(slot) => {
-                let target = self.target_slot(SlotId(slot));
-                self.open_indicator_settings_at(target);
-            }
-            // The toolbar acts on the market it is showing: the active tab's
-            // simulator, whose tape the buttons' price came from.
-            ToolbarAction::PaperBuy => self
-                .active_tab_mut()
-                .paper
-                .market(quantick_engine::Side::Buy),
-            ToolbarAction::PaperSell => self
-                .active_tab_mut()
-                .paper
-                .market(quantick_engine::Side::Sell),
-            ToolbarAction::PaperClose => self.active_tab_mut().paper.close_position(),
-        }
-    }
-
     /// Hand the app the window it is drawing into.
     ///
     /// Called once from `main`, which is where eframe offers the handle: the
@@ -2638,1209 +1702,6 @@ impl QuantickApp {
     /// to, which is also what every non-Windows target gets.
     pub fn attach_surface(&mut self, handle: &impl raw_window_handle::HasWindowHandle) {
         self.surface = window_scale::SurfaceProbe::new(handle);
-    }
-}
-
-/// Opens the Market Replay browser (§10).
-const REPLAY_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::R);
-/// Shows/hides the panels dock (§10).
-/// `Ctrl+1` … `Ctrl+9` apply the layout registry's presets, in table order.
-///
-/// The keys are listed; which preset each one reaches is not. A preset added
-/// to `LAYOUT_PRESETS` gets its shortcut from its position without this array
-/// or its dispatch being edited — the same rule the picker and the View menu
-/// follow. Nine is what a number row has; `MAX_CANVAS_PANES` keeps the
-/// registry far below that.
-const LAYOUT_PRESET_KEYS: [egui::Key; 9] = [
-    egui::Key::Num1,
-    egui::Key::Num2,
-    egui::Key::Num3,
-    egui::Key::Num4,
-    egui::Key::Num5,
-    egui::Key::Num6,
-    egui::Key::Num7,
-    egui::Key::Num8,
-    egui::Key::Num9,
-];
-
-/// The shortcut that reaches the preset at `index` in the registry, if a
-/// number key still reaches that far.
-fn layout_preset_shortcut(index: usize) -> Option<egui::KeyboardShortcut> {
-    LAYOUT_PRESET_KEYS
-        .get(index)
-        .map(|key| egui::KeyboardShortcut::new(egui::Modifiers::CTRL, *key))
-}
-
-/// The shortcut that reaches the layout tab at strip position `index`, if a
-/// number key still reaches that far.
-fn layout_tab_shortcut(index: usize) -> Option<egui::KeyboardShortcut> {
-    LAYOUT_PRESET_KEYS
-        .get(index)
-        .map(|key| egui::KeyboardShortcut::new(egui::Modifiers::ALT, *key))
-}
-
-/// `Ctrl+0` puts the context charts away, or brings them back.
-///
-/// The number row's own zero, beside `Ctrl+1..9` for the presets: nine keys
-/// choose an arrangement and the tenth dismisses the column that arrangement
-/// put beside the heatmap. Without it the only way to collapse was a drag,
-/// which a trader working by keyboard cannot make and which WCAG 2.2's
-/// dragging rule wants an alternative to besides.
-const COLLAPSE_CONTEXT_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::Num0);
-
-const DOCK_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::B);
-/// Folds the focused pane's on-chart indicator legend to its count puck, or
-/// opens it back up (see [`crate::indicator_legend`]).
-///
-/// Ctrl+letter like the dock's own switch above, not the bare `L` the drawing
-/// tools answer to: bare letters are the toolbox's namespace, and a chrome
-/// switch borrowing one would arm a tool on every trader who learned it there.
-const LEGEND_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::L);
-/// Saves the workspace — the arrangement the next launch opens on.
-///
-/// Ctrl+Shift+S rather than the Ctrl+S every editor uses, deliberately: a
-/// chart has no document, and a trader who reaches for Ctrl+S out of habit
-/// mid-session should hit nothing rather than silently redefine what their
-/// platform opens on.
-const SAVE_WORKSPACE_SHORTCUT: egui::KeyboardShortcut = egui::KeyboardShortcut::new(
-    egui::Modifiers::CTRL.plus(egui::Modifiers::SHIFT),
-    egui::Key::S,
-);
-/// Opens the source picker for a new tab (§10).
-const NEW_TAB_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::T);
-/// Closes the active tab (§10). Free of any other binding: the chart has no
-/// text inputs and no document to "write".
-const CLOSE_TAB_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::W);
-/// Cycles forward through the strip (§10).
-const NEXT_TAB_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::CTRL, egui::Key::Tab);
-/// See [`NEXT_TAB_SHORTCUT`].
-const PREVIOUS_TAB_SHORTCUT: egui::KeyboardShortcut = egui::KeyboardShortcut::new(
-    egui::Modifiers::CTRL.plus(egui::Modifiers::SHIFT),
-    egui::Key::Tab,
-);
-/// Simulated buy at market (`docs/ux/paper-trading.md` §9). All the
-/// trading hotkeys are Shift+letter and stand down while any text field
-/// owns the keyboard — a capital letter typed into a symbol box must
-/// never become an order.
-const PAPER_BUY_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::B);
-/// Simulated sell at market.
-const PAPER_SELL_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::S);
-/// Reverse the simulated position.
-const PAPER_REVERSE_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::R);
-/// Flatten: close the position and cancel every working order.
-const PAPER_FLATTEN_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::F);
-/// Cancel every working order without trading.
-const PAPER_CANCEL_SHORTCUT: egui::KeyboardShortcut =
-    egui::KeyboardShortcut::new(egui::Modifiers::SHIFT, egui::Key::X);
-/// Height of the menu bar, in pixels (§5 zone 1).
-const MENU_BAR_HEIGHT: f32 = 28.0;
-
-impl QuantickApp {
-    /// The window's menu bar (§10): shallow menus for discoverability and
-    /// shortcuts, never the only path to anything.
-    fn draw_menu_bar(&mut self, ctx: &egui::Context) {
-        if ctx.input_mut(|i| i.consume_shortcut(&REPLAY_SHORTCUT)) {
-            self.replay_view.open_browser();
-        }
-        if ctx.input_mut(|i| i.consume_shortcut(&DOCK_SHORTCUT)) {
-            self.dock.toggle_visible();
-        }
-        // Gated on the same condition the View menu's entry is gated on: only
-        // a layout that carves a context column *beside* the flow pane has a
-        // column to put away. Ungated, `Ctrl+0` on the Flow or Timeframe
-        // layout set a flag nothing drew — and swallowed the key besides, so
-        // egui's own "reset zoom" never ran.
-        if self.active_tab().layout.shows_time()
-            && self.active_tab().layout.shows_flow()
-            && ctx.input_mut(|i| i.consume_shortcut(&COLLAPSE_CONTEXT_SHORTCUT))
-        {
-            let collapsed = self.active_tab().context_collapsed;
-            self.active_tab_mut().set_context_collapsed(!collapsed);
-        }
-        // Layout by number, straight off the registry. The same
-        // `apply_layout_preset` the picker and the menu call — three doors,
-        // one room.
-        for (index, preset) in crate::canvas_layout::LAYOUT_PRESETS.iter().enumerate() {
-            let Some(shortcut) = layout_preset_shortcut(index) else {
-                break;
-            };
-            if ctx.input_mut(|i| i.consume_shortcut(&shortcut)) {
-                self.apply_layout_preset(preset);
-            }
-        }
-        // Layout tabs by number: `Alt+1..9`, beside `Ctrl+1..9` for the
-        // presets — one row of keys, two things a trader switches by number.
-        // Not while a text field has the keyboard — a rename box, a note,
-        // the ticket — where Alt+1 is text, not a switch.
-        let typing = ctx.memory(|memory| memory.focused().is_some());
-        for index in 0..LAYOUT_PRESET_KEYS.len() {
-            let Some(shortcut) = layout_tab_shortcut(index) else {
-                break;
-            };
-            if !typing
-                && ctx.input_mut(|i| i.consume_shortcut(&shortcut))
-                && let Err(error) = self.switch_layout_index(index)
-            {
-                self.note_workspace(error.to_string());
-            }
-        }
-        if ctx.input_mut(|i| i.consume_shortcut(&LEGEND_SHORTCUT)) {
-            let collapsed = self.focused_legend_collapsed();
-            self.set_focused_legend_collapsed(!collapsed);
-        }
-        if ctx.input_mut(|i| i.consume_shortcut(&crate::control::MARK_SHORTCUT)) {
-            self.take_mark(None);
-        }
-        if ctx.input_mut(|i| i.consume_shortcut(&SAVE_WORKSPACE_SHORTCUT)) {
-            self.save_workspace("shortcut");
-        }
-        // Trading hotkeys, swallowed only while no text field owns the
-        // keyboard. Market entries use the ticket's quantity and offsets,
-        // exactly like the toolbar buttons they twin.
-        if !ctx.wants_keyboard_input() {
-            if ctx.input_mut(|i| i.consume_shortcut(&PAPER_BUY_SHORTCUT)) {
-                self.active_tab_mut()
-                    .paper
-                    .market(quantick_engine::Side::Buy);
-            }
-            if ctx.input_mut(|i| i.consume_shortcut(&PAPER_SELL_SHORTCUT)) {
-                self.active_tab_mut()
-                    .paper
-                    .market(quantick_engine::Side::Sell);
-            }
-            if ctx.input_mut(|i| i.consume_shortcut(&PAPER_REVERSE_SHORTCUT)) {
-                self.active_tab_mut().paper.reverse_position();
-            }
-            if ctx.input_mut(|i| i.consume_shortcut(&PAPER_FLATTEN_SHORTCUT)) {
-                self.active_tab_mut().paper.flatten();
-            }
-            if ctx.input_mut(|i| i.consume_shortcut(&PAPER_CANCEL_SHORTCUT)) {
-                self.active_tab_mut()
-                    .paper
-                    .account_mut()
-                    .cancel_all_orders();
-            }
-        }
-
-        let mut tab_action = None;
-        egui::TopBottomPanel::top("menu_bar")
-            .exact_height(MENU_BAR_HEIGHT)
-            .frame(
-                egui::Frame::none()
-                    .fill(theme::CHROME)
-                    .inner_margin(egui::Margin::symmetric(6.0, 4.0)),
-            )
-            .show(ctx, |ui| {
-                egui::menu::bar(ui, |ui| {
-                    ui.menu_button("File", |ui| {
-                        if ui
-                            .add(
-                                egui::Button::new("New Tab…")
-                                    .shortcut_text(ui.ctx().format_shortcut(&NEW_TAB_SHORTCUT)),
-                            )
-                            .clicked()
-                        {
-                            tab_action = Some(TabAction::New);
-                            ui.close_menu();
-                        }
-                        if ui
-                            .add_enabled(
-                                self.tabs.len() > 1,
-                                egui::Button::new("Close Tab").shortcut_text(
-                                    ui.ctx().format_shortcut(&CLOSE_TAB_SHORTCUT),
-                                ),
-                            )
-                            .on_disabled_hover_text(
-                                "The last tab stays open — a window with no market has nothing to show",
-                            )
-                            .clicked()
-                        {
-                            tab_action = Some(TabAction::Close(self.active_tab));
-                            ui.close_menu();
-                        }
-                        ui.separator();
-                        if ui
-                            .add(
-                                egui::Button::new("Market Replay…")
-                                    .shortcut_text(ui.ctx().format_shortcut(&REPLAY_SHORTCUT)),
-                            )
-                            .clicked()
-                        {
-                            self.replay_view.open_browser();
-                            ui.close_menu();
-                        }
-                        if self.active_tab().replay.is_some() && ui.button("Close Replay").clicked()
-                        {
-                            let (tab, config) = self.active_with_config();
-                            tab.close_replay(config);
-                            ui.close_menu();
-                        }
-                        ui.separator();
-                        if ui.button("Exit").clicked() {
-                            ui.ctx().send_viewport_cmd(egui::ViewportCommand::Close);
-                        }
-                    });
-                    ui.menu_button("View", |ui| {
-                        // What the canvas shows is a view concern, so the
-                        // switch lives here rather than under File, and each
-                        // entry names the charts it shows — "Timeframe", not
-                        // layout jargon (audit §3).
-                        ui.menu_button("Layouts", |ui| {
-                            // The strip's tabs, from the book: switch the
-                            // focused pane by name, and the three edits the
-                            // strip's own menu holds.
-                            let active = self.focused_pane_layout();
-                            let names: Vec<(crate::layouts::LayoutId, String)> = self
-                                .layouts()
-                                .layouts()
-                                .iter()
-                                .map(|layout| (layout.id, layout.name.clone()))
-                                .collect();
-                            for (index, (id, name)) in names.iter().enumerate() {
-                                let mut button = egui::Button::new(name.as_str());
-                                if let Some(shortcut) = layout_tab_shortcut(index) {
-                                    button = button.shortcut_text(ui.ctx().format_shortcut(&shortcut));
-                                }
-                                if ui.add(button.selected(*id == active)).clicked() {
-                                    self.apply_strip_action(crate::layout_strip::StripAction::Switch(*id));
-                                    ui.close_menu();
-                                }
-                            }
-                            ui.separator();
-                            let can_add = names.len() < crate::layouts::MAX_LAYOUTS;
-                            if ui
-                                .add_enabled(can_add, egui::Button::new("New layout"))
-                                .clicked()
-                            {
-                                self.apply_strip_action(crate::layout_strip::StripAction::Create);
-                                ui.close_menu();
-                            }
-                            if ui.button("Rename layout…").clicked() {
-                                self.apply_strip_action(crate::layout_strip::StripAction::BeginRename(active));
-                                ui.close_menu();
-                            }
-                            if ui
-                                .add_enabled(names.len() > 1, egui::Button::new("Delete layout"))
-                                .clicked()
-                            {
-                                self.apply_strip_action(crate::layout_strip::StripAction::Delete(active));
-                                ui.close_menu();
-                            }
-                        });
-                        ui.menu_button("Layout", |ui| {
-                            // Read from the registry, like the picker: a menu
-                            // holding its own list of layouts is the second
-                            // opinion that goes stale the day one is added.
-                            let current = self.active_tab().layout.preset();
-                            for (index, preset) in
-                                crate::canvas_layout::LAYOUT_PRESETS.iter().enumerate()
-                            {
-                                // The menu is where a shortcut is learned, so
-                                // it carries the binding beside the name.
-                                let label = match layout_preset_shortcut(index) {
-                                    Some(shortcut) => format!(
-                                        "{}	{}",
-                                        preset.label,
-                                        ui.ctx().format_shortcut(&shortcut)
-                                    ),
-                                    None => preset.label.to_owned(),
-                                };
-                                if ui
-                                    .selectable_label(current.id == preset.id, label)
-                                    .clicked()
-                                {
-                                    self.apply_layout_preset(preset);
-                                    ui.close_menu();
-                                }
-                            }
-                        });
-                        // Collapsing was drag-only, which a trader working
-                        // by keyboard could not do at all — while the
-                        // assistant had `layout.pane.collapse`. Same call,
-                        // three doors.
-                        if self.active_tab().layout.shows_time()
-                            && self.active_tab().layout.shows_flow()
-                        {
-                            let collapsed = self.active_tab().context_collapsed;
-                            let label = if collapsed {
-                                "Show context charts"
-                            } else {
-                                "Hide context charts"
-                            };
-                            if ui
-                                .add(
-                                    egui::Button::new(label).shortcut_text(
-                                        ui.ctx().format_shortcut(&COLLAPSE_CONTEXT_SHORTCUT),
-                                    ),
-                                )
-                                .clicked()
-                            {
-                                self.active_tab_mut().set_context_collapsed(!collapsed);
-                                ui.close_menu();
-                            }
-                        }
-                        // Reposition without a drag. WCAG 2.2's dragging rule
-                        // wants a single-pointer alternative to every drag, and
-                        // TradingView — the reference the trader named — moves
-                        // charts by a menu command rather than by dragging at
-                        // all. Both go through `Tab::move_context_pane`.
-                        let context_panes = self.active_tab().pane_count().saturating_sub(1);
-                        if context_panes > 1 {
-                            ui.menu_button("Move chart", |ui| {
-                                for slot in 1..=context_panes {
-                                    let up = ui
-                                        .add_enabled(slot > 1, egui::Button::new(format!(
-                                            "Chart {slot} up"
-                                        )))
-                                        .on_disabled_hover_text("already the top chart");
-                                    if up.clicked() {
-                                        let tab_id = self.active_tab().id;
-                                        self.move_context_pane_at(tab_id, slot, slot - 1);
-                                        ui.close_menu();
-                                    }
-                                    let down = ui
-                                        .add_enabled(
-                                            slot < context_panes,
-                                            egui::Button::new(format!("Chart {slot} down")),
-                                        )
-                                        .on_disabled_hover_text("already the bottom chart");
-                                    if down.clicked() {
-                                        let tab_id = self.active_tab().id;
-                                        self.move_context_pane_at(tab_id, slot, slot + 1);
-                                        ui.close_menu();
-                                    }
-                                }
-                            });
-                        }
-                        ui.separator();
-                        let panels_label = if self.dock.visible() {
-                            "Hide panels"
-                        } else {
-                            "Show panels"
-                        };
-                        if ui
-                            .add(
-                                egui::Button::new(panels_label)
-                                    .shortcut_text(ui.ctx().format_shortcut(&DOCK_SHORTCUT)),
-                            )
-                            .clicked()
-                        {
-                            self.dock.toggle_visible();
-                            ui.close_menu();
-                        }
-                        // The legend belongs to a pane, so this entry names
-                        // the focused one's state — the same pane the chevron
-                        // on screen would fold.
-                        let collapsed = self.focused_legend_collapsed();
-                        // Split open: say *which* chart, the way the layout
-                        // entries above name the charts they show. The action
-                        // follows the focus like every other chrome control,
-                        // and a trader reading "Collapse indicator legend"
-                        // over two charts has no way to know which corner is
-                        // about to change.
-                        let split = self.active_tab().shows_context_charts();
-                        let pane_name = self.active_tab().focused_side().title();
-                        let legend_label = match (collapsed, split) {
-                            (true, false) => "Show indicator legend".to_owned(),
-                            (false, false) => "Collapse indicator legend".to_owned(),
-                            (true, true) => format!("Show indicator legend ({pane_name})"),
-                            (false, true) => format!("Collapse indicator legend ({pane_name})"),
-                        };
-                        // A pane with no indicators has no legend to fold, and
-                        // an entry that is enabled and does nothing reads as a
-                        // broken feature rather than as an empty chart.
-                        let has_legend = {
-                            let tab = self.active_tab();
-                            !tab.pane(tab.focused_side()).indicators.all().is_empty()
-                        };
-                        if ui
-                            .add_enabled(
-                                has_legend,
-                                egui::Button::new(legend_label)
-                                    .shortcut_text(ui.ctx().format_shortcut(&LEGEND_SHORTCUT)),
-                            )
-                            .on_hover_text(
-                                "Folds the healthy rows to a count on the focused chart. Errored and stale indicators stay on it.",
-                            )
-                            .on_disabled_hover_text(
-                                "This chart has no indicators, so there is no legend to fold",
-                            )
-                            .clicked()
-                        {
-                            self.set_focused_legend_collapsed(!collapsed);
-                            ui.close_menu();
-                        }
-                        ui.menu_button("Drawing toolbar", |ui| {
-                            for (dock, label) in [
-                                (ToolboxDock::Left, "Left"),
-                                (ToolboxDock::Top, "Top"),
-                                (ToolboxDock::Bottom, "Bottom"),
-                            ] {
-                                if ui
-                                    .selectable_label(self.toolrail.dock() == dock, label)
-                                    .clicked()
-                                {
-                                    self.toolrail.set_dock(dock);
-                                    ui.close_menu();
-                                }
-                            }
-                        });
-                        let toolbox_label = if self.toolrail.visible() {
-                            "Hide drawing toolbar"
-                        } else {
-                            "Show drawing toolbar"
-                        };
-                        if ui.button(toolbox_label).clicked() {
-                            self.toolrail.toggle_visible();
-                            ui.close_menu();
-                        }
-                        for (tab, label) in [
-                            (DockTab::L2, "L2 settings"),
-                            (DockTab::Bubbles, "Bubble settings"),
-                            (DockTab::Session, "Session"),
-                            (DockTab::Trading, "Paper trading"),
-                            (DockTab::Trades, "Trades"),
-                        ] {
-                            if ui.button(label).clicked() {
-                                self.dock.open_tab(tab);
-                                ui.close_menu();
-                            }
-                        }
-                        ui.separator();
-                        ui.checkbox(&mut self.show_perf, "Perf readings")
-                            .on_hover_text("fps, frame time and trade count on the status bar");
-                        ui.checkbox(&mut self.progressive_history, "Progressive venue history")
-                            .on_hover_text(
-                                "Build the venue's candle history from now backwards, a week at \
-                                 a time, so the chart fills in while the rest arrives. Off asks \
-                                 for the whole span in one request: fewer calls, nothing on \
-                                 screen until all of it lands.",
-                            );
-                        ui.checkbox(&mut self.venue_lead_in, "Venue candles on charts cut by trades")
-                            .on_hover_text(
-                                "A tick, volume, dollar or imbalance chart cannot fold venue \
-                                 candles into its own bars, so it opens holding only the prints \
-                                 this session saw. Switch this on to put the venue's 1-minute \
-                                 candles in front of them anyway — counted apart from built bars \
-                                 on the status bar — so yesterday is on screen to compare \
-                                 against. They stay candles: a minute never becomes a tick bar, \
-                                 and an indicator running across the seam is averaging both \
-                                 kinds.",
-                            );
-                        ui.separator();
-                        ui.menu_button("Timezone", |ui| {
-                            egui::ScrollArea::vertical()
-                                .max_height(280.0)
-                                .show(ui, |ui| {
-                                    for tz in TzOffset::ALL {
-                                        if ui.selectable_label(self.tz == tz, tz.label()).clicked()
-                                        {
-                                            self.tz = tz;
-                                            ui.close_menu();
-                                        }
-                                    }
-                                });
-                        });
-                    });
-                    // The workspace is its own menu, not a File entry: "what
-                    // does quantick open on" is a question a trader asks about
-                    // their cockpit, not about a document, and burying it
-                    // under File is how a platform ends up with traders who
-                    // rebuild their screen every morning without knowing they
-                    // never had to (audit §6).
-                    // The button's own rect, published for the capture hook —
-                    // read, never acted on, so the menu behaves identically
-                    // whether or not a scripted run is watching.
-                    let workspace_menu = ui.menu_button("Workspace", |ui| {
-                        if ui
-                            .add(
-                                egui::Button::new("Save workspace").shortcut_text(
-                                    ui.ctx().format_shortcut(&SAVE_WORKSPACE_SHORTCUT),
-                                ),
-                            )
-                            .on_hover_text(
-                                "Remember this arrangement — the tabs, the charts on each, the \
-                                 panels, the timezone and the window — as what quantick opens on",
-                            )
-                            .clicked()
-                        {
-                            self.save_workspace("menu");
-                            ui.close_menu();
-                        }
-                        // Enabled only when there is something on disk to go
-                        // back to: an entry that would forget nothing is a
-                        // question the trader should not have to answer by
-                        // clicking it.
-                        if ui
-                            .add_enabled(
-                                self.workspace.session().saved(),
-                                egui::Button::new("Reset startup layout"),
-                            )
-                            .on_hover_text(
-                                "Forget the saved workspace; the next launch opens on the \
-                                 configured default. The charts on screen are left alone.",
-                            )
-                            .on_disabled_hover_text(
-                                "Nothing saved yet — quantick already opens on the configured \
-                                 default",
-                            )
-                            .clicked()
-                        {
-                            self.forget_workspace();
-                            ui.close_menu();
-                        }
-                        ui.separator();
-                        // Bookmarks. Named apart from the two entries above on
-                        // purpose: those govern what the app *opens on*, these
-                        // are places to come back to. The wording carries the
-                        // distinction so the menu does not need a paragraph.
-                        if ui
-                            .button("Save as…")
-                            // Says what a bookmark keeps *and what it does
-                            // not*: it is the tabs and the panels, not the
-                            // indicators or the colours. Two entries in one
-                            // menu that both "save a workspace" but restore
-                            // different amounts is exactly how a trader comes
-                            // to believe the app forgets things — use Export
-                            // to file for the whole cockpit.
-                            .on_hover_text(
-                                "Keep these tabs and panels under a name you can reopen later. \
-                                 It does not change what quantick opens on, and it does not \
-                                 keep indicators or colours — use Export to file for those.",
-                            )
-                            .clicked()
-                        {
-                            self.surfaces.workspace_name.open();
-                            ui.close_menu();
-                        }
-                        let mut open: Option<String> = None;
-                        let mut delete: Option<String> = None;
-                        ui.add_enabled_ui(!self.workspace.session().bookmarks().is_empty(), |ui| {
-                            ui.menu_button("Open", |ui| {
-                                for entry in self.workspace.session().bookmarks() {
-                                    let tabs = entry.tabs.len();
-                                    if ui
-                                        .button(&entry.name)
-                                        .on_hover_text(format!(
-                                            "{tabs} chart {} — replaces what is on screen",
-                                            if tabs == 1 { "tab" } else { "tabs" }
-                                        ))
-                                        .clicked()
-                                    {
-                                        open = Some(entry.name.clone());
-                                        ui.close_menu();
-                                    }
-                                }
-                            })
-                            .response
-                            .on_disabled_hover_text("Nothing saved under a name yet");
-                            ui.menu_button("Delete", |ui| {
-                                for entry in self.workspace.session().bookmarks() {
-                                    if ui.button(&entry.name).clicked() {
-                                        delete = Some(entry.name.clone());
-                                        ui.close_menu();
-                                    }
-                                }
-                            });
-                        });
-                        if let Some(name) = open {
-                            self.open_named_workspace(&name);
-                        }
-                        if let Some(name) = delete {
-                            self.delete_named_workspace(&name);
-                        }
-                        ui.separator();
-                        // Files, named apart from the two groups above again:
-                        // those live inside quantick, these are documents the
-                        // trader owns, can copy, back up and carry to another
-                        // machine. That is the difference the wording carries.
-                        if ui
-                            .button("Export to file…")
-                            .on_hover_text(
-                                "Save the whole cockpit — tabs, indicators, layers, drawing \
-                                 colours, footprint and added symbols — as one file in your \
-                                 documents",
-                            )
-                            .clicked()
-                        {
-                            self.open_workspace_export_picker();
-                            ui.close_menu();
-                        }
-                        if ui
-                            .button("Open from file…")
-                            .on_hover_text(
-                                "Open a workspace file. It replaces the cockpit on screen; a \
-                                 file that cannot be read changes nothing.",
-                            )
-                            .clicked()
-                        {
-                            self.open_workspace_import_picker();
-                            ui.close_menu();
-                        }
-                        // Read off the field, not the filesystem: this body
-                        // runs every frame the menu is open.
-                        let mut reopen: Option<std::path::PathBuf> = None;
-                        ui.add_enabled_ui(!self.workspace.session().recent_on_disk().is_empty(), |ui| {
-                            ui.menu_button("Open recent", |ui| {
-                                for path in self.workspace.session().recent_on_disk() {
-                                    if ui
-                                        .button(crate::workspace_bundle::recent_label(path))
-                                        // The same warning the bookmark list
-                                        // carries: this replaces the cockpit,
-                                        // and a trader mid-tape has to read
-                                        // that before the click, not after.
-                                        .on_hover_text(format!(
-                                            "Replaces the cockpit on screen\n{}",
-                                            path.display()
-                                        ))
-                                        .clicked()
-                                    {
-                                        reopen = Some(path.clone());
-                                        ui.close_menu();
-                                    }
-                                }
-                            })
-                            .response
-                            .on_disabled_hover_text("No workspace files opened yet");
-                        });
-                        if let Some(path) = reopen {
-                            self.import_workspace_from(&path);
-                        }
-                        if ui
-                            .button("Show where it's saved")
-                            .on_hover_text(
-                                "Open the folder quantick keeps your cockpit in, so you can see \
-                                 it and back it up",
-                            )
-                            .clicked()
-                        {
-                            self.reveal_cockpit_home();
-                            ui.close_menu();
-                        }
-                        ui.separator();
-                        if ui
-                            .checkbox(self.workspace.session_mut().save_on_exit_mut(), "Save on exit")
-                            .on_hover_text(
-                                "Keep the arrangement automatically when the window closes. Off, \
-                                 only Save workspace changes what quantick opens on.",
-                            )
-                            .changed()
-                        {
-                            // The setting lives in the file it governs, so
-                            // switching it has to reach the disk now — not at
-                            // the next exit, which is exactly the exit it may
-                            // have just switched off.
-                            self.save_workspace("save_on_exit_toggled");
-                        }
-                    });
-                    self.workspace_menu_rect = Some(workspace_menu.response.rect);
-                    ui.menu_button("Tools", |ui| {
-                        if ui.button("Appearance…").clicked() {
-                            self.surfaces.style_panel.open();
-                            ui.close_menu();
-                        }
-                        let access_label = self
-                            .control_access
-                            .as_ref()
-                            .map_or("Local agent access…", |access| access.menu_label());
-                        if ui.button(access_label).clicked() {
-                            if let Some(access) = self.control_access.as_mut() {
-                                access.open_panel();
-                            }
-                            ui.close_menu();
-                        }
-                    });
-                    ui.menu_button("Help", |ui| {
-                        if ui.button("Replay file format…").clicked() {
-                            self.replay_view.open_format_help();
-                            ui.close_menu();
-                        }
-                    });
-                    if self
-                        .control_access
-                        .as_ref()
-                        .is_some_and(crate::control::ControlAccess::is_enabled)
-                        && ui.button("Agent access: on").clicked()
-                        && let Some(access) = self.control_access.as_mut()
-                    {
-                        access.open_panel();
-                    }
-                    ui.separator();
-                    // The tab strip shares the menu row: zone 1 already had
-                    // the horizontal room, so tabs cost no chrome budget.
-                    tab_action = self.draw_tab_strip(ui);
-                });
-            });
-        if let Some(action) = tab_action {
-            self.apply_tab_action(action);
-        }
-    }
-
-    /// The chips, built from what each tab actually is right now.
-    fn draw_tab_strip(&self, ui: &mut egui::Ui) -> Option<TabAction> {
-        let chips: Vec<tabstrip::TabChip<'_>> = self
-            .tabs
-            .iter()
-            .map(|tab| tabstrip::TabChip {
-                label: tab.chip_label(),
-                replaying: tab.replay.is_some(),
-                needs_attention: tab.needs_attention(),
-            })
-            .collect();
-        tabstrip::draw(ui, &chips, self.active_tab)
-    }
-
-    /// One delete command for every trigger (inspector button, keyboard,
-    /// manager). A locked object raises the confirmation next to the trigger
-    /// instead of deleting; a landed delete raises the Undo toast.
-    fn request_delete_selected(&mut self, now: Instant) {
-        // Read the name before the object is gone. "Drawing deleted" makes
-        // the undo useless on a crowded chart: the trader has to know *what*
-        // they lost to know whether they want it back — and the context bar
-        // deletes on a bare glyph, so the toast is what pays for that.
-        let doomed = self.drawing_pane().drawings.selected().and_then(|index| {
-            let drawing = self.drawing_pane().drawings.items().get(index)?;
-            // The trader's own name when one was given; the tool name
-            // otherwise — a positional index would be noise on an object
-            // that no longer has a position.
-            let label = drawing
-                .name
-                .clone()
-                .unwrap_or_else(|| drawing.tool.name().to_owned());
-            Some((drawing.id, label))
-        });
-        match self.drawing_pane_mut().drawings.delete_selected(false) {
-            DeleteOutcome::Deleted => {
-                self.surfaces.drawing_chrome.set_delete_confirm(false);
-                // The instance dies with its drawing, immediately — not on
-                // the next closed bar, which a quiet tape may never bring.
-                if let Some((id, _)) = &doomed {
-                    self.drawing_pane_mut().remove_strategy_for_drawing(*id);
-                }
-                let name = doomed.map(|(_, label)| label);
-                let message = name.map_or_else(
-                    || "Drawing deleted.".to_owned(),
-                    |name| format!("{name} deleted."),
-                );
-                self.surfaces.toast.note_with_undo(message, now);
-            }
-            DeleteOutcome::NeedsConfirmation => {
-                self.surfaces.drawing_chrome.set_delete_confirm(true);
-            }
-            DeleteOutcome::NothingSelected => {}
-        }
-    }
-
-    /// Record one edit gesture as the single undo entry it earned, on the pane
-    /// it started on.
-    ///
-    /// That pane's tab may have been closed under the gesture, in which case
-    /// the object it described is gone with it.
-    fn record_drawing_edit(
-        &mut self,
-        tab_id: u64,
-        side: PaneSide,
-        index: usize,
-        before: drawings::Drawing,
-    ) {
-        if let Some(tab) = self.tabs.iter_mut().find(|tab| tab.id == tab_id) {
-            tab.pane_mut(side).drawings.record_edit_of(index, before);
-        }
-    }
-
-    /// Put the caret in a note, on the chart — the one call that opens the
-    /// editor, whether a placement, a double click or a script asked for it.
-    ///
-    /// The surface decides whether the caret is allowed: an object that holds
-    /// no words, or a locked one, refuses it, because an editor that opened and
-    /// then dropped every keystroke would be worse than none. The store command
-    /// and the per-pane stand-down are the host's, so they happen here.
-    pub fn begin_inline_text_edit(&mut self, index: usize) -> bool {
-        let Self {
-            tabs,
-            active_tab,
-            surfaces,
-            ..
-        } = self;
-        let tab = &tabs[*active_tab];
-        let side = tab.drawing_side();
-        let Some(drawing) = tab.pane(side).drawings.items().get(index) else {
-            return false;
-        };
-        if !surfaces
-            .drawing_chrome
-            .begin_inline_text_edit(tab.id, side, index, drawing)
-        {
-            return false;
-        }
-        self.drawing_pane_mut().drawings.select(Some(index));
-        self.sync_content_editing();
-        true
-    }
-
-    /// Close the editor, keeping whatever was typed and recording it as the one
-    /// edit it was — on the pane the note actually lives on, which is not
-    /// necessarily the one in front when it closes.
-    fn end_inline_text_edit(&mut self) {
-        if let Some(edit) = self.surfaces.drawing_chrome.end_inline_text_edit() {
-            self.record_drawing_edit(edit.tab, edit.side, edit.index, edit.before);
-        }
-        self.sync_content_editing();
-    }
-
-    /// Tell every pane whether one of its objects is having its content typed
-    /// somewhere else on screen, so exactly one object anywhere stands down.
-    ///
-    /// Every pane, not just the one in front: the flag is what suppresses the
-    /// object's own painting, and a pane left holding a stale index would keep
-    /// a note invisible for the rest of the session with no way back.
-    fn sync_content_editing(&mut self) {
-        let editing = self.surfaces.drawing_chrome.content_editing_target();
-        for tab in &mut self.tabs {
-            let target = editing
-                .filter(|(id, _, _)| *id == tab.id)
-                .map(|(_, side, index)| (side, index));
-            tab.set_content_editing(target);
-        }
-    }
-
-    /// Which note is being typed on the chart right now — what a second
-    /// operator reads to know the keyboard belongs to an object.
-    #[must_use]
-    pub fn inline_text_editing(&self) -> Option<usize> {
-        self.surfaces.drawing_chrome.inline_text_editing()
-    }
-
-    /// The rows the object manager lists.
-    ///
-    /// A row's facts come from the drawing, the pane's band registry and the
-    /// tab's layout, and assembling them here is what keeps the manager from
-    /// needing all three. Built only while the window is open, like the market
-    /// dialog's list of open markets: a dozen short strings once a frame, on a
-    /// window that is shut the rest of the session.
-    fn drawing_manager_rows(&self) -> Vec<crate::surfaces::drawing_chrome::ManagerRow> {
-        let pane = self.drawing_pane();
-        let selected = pane.drawings.selected();
-        let focused = self.focused_pane();
-        pane.drawings
-            .items()
-            .iter()
-            .enumerate()
-            .map(
-                |(index, drawing)| crate::surfaces::drawing_chrome::ManagerRow {
-                    name: drawing.display_label(index),
-                    selected: selected == Some(index),
-                    locked: drawing.locked,
-                    hidden: drawing.hidden,
-                    shared: drawing.scope == drawings::DrawingScope::AllCharts,
-                    off_series: drawing.off_series,
-                    foreign_market: drawing.foreign_market,
-                    author: drawing.author.as_ref().map(DrawingAuthor::label),
-                    band: focused.band_label(drawing),
-                },
-            )
-            .collect()
-    }
-
-    /// Where the selected object is painted, in screen points. The chrome
-    /// cannot work this out for itself: it needs the viewport and the price
-    /// scale the host owns.
-    ///
-    /// Two separate answers rather than one pair, because they cost different
-    /// things and not every pass wants both — this one walks the object's
-    /// anchors through the price scale. Nothing is projected while nothing is
-    /// selected, which is every frame of an ordinary session.
-    fn selected_drawing_bbox(&self) -> Option<egui::Rect> {
-        let pane = self.drawing_pane();
-        let index = pane.drawings.selected()?;
-        let chart = pane.last_chart_area?;
-        self.drawing_bbox_on_screen(chart, index)
-    }
-
-    /// What the band the selected object lives on is called, for the
-    /// inspector's title. `None` on the price band, where a suffix on every
-    /// object would be noise. Formats a string, so it is asked for only by a
-    /// pass that shows the title.
-    fn selected_drawing_band(&self) -> Option<String> {
-        let pane = self.drawing_pane();
-        let index = pane.drawings.selected()?;
-        self.focused_pane()
-            .band_label(pane.drawings.items().get(index)?)
-            .chip()
-    }
-
-    /// The docked inspector.
-    ///
-    /// Its own call site because a `SidePanel` has to be declared *before* the
-    /// central canvas — the canvas pays its width, and a panel declared after
-    /// it would overlay the chart instead of docking beside it.
-    fn draw_pinned_inspector(&mut self, ctx: &egui::Context, now: Instant) {
-        if !self.surfaces.drawing_chrome.inspector_pinned() {
-            return;
-        }
-        // No painted bounds: a docked panel has no placement rule to keep
-        // clear of the object, so the projection the floating one needs is not
-        // gathered here. The band still is — it is in the title.
-        let read = DrawingRead {
-            selected_band: self.selected_drawing_band(),
-            ..DrawingRead::default()
-        };
-        let ask = self.draw_chrome_pass(ctx, read, false);
-        self.apply_drawing_chrome(ask, now);
-    }
-
-    /// The four floating pieces, registered after the canvas so they stay in
-    /// front of the chart they are anchored to.
-    fn draw_drawing_chrome(&mut self, ctx: &egui::Context, now: Instant) {
-        let manager_open = self.surfaces.drawing_chrome.manager_open();
-        let rows = if manager_open {
-            self.drawing_manager_rows()
-        } else {
-            Vec::new()
-        };
-        // The band name goes in the inspector's title and nowhere else, so
-        // it is formatted only when one of the two inspector hosts is on
-        // screen. A selection alone raises the context bar, which never shows
-        // it — and `band_label` scans the pane's indicator views and `chip`
-        // allocates, every frame, for a value nothing would read.
-        let inspector_showing = self.surfaces.drawing_chrome.inspector_open()
-            || self.surfaces.drawing_chrome.inspector_pinned();
-        let read = DrawingRead {
-            selected_bbox: self.selected_drawing_bbox(),
-            selected_band: inspector_showing
-                .then(|| self.selected_drawing_band())
-                .flatten(),
-            // Counted only for the window that offers to take them back, and
-            // over every tab: an object an assistant placed on another chart
-            // still belongs in that count.
-            authored_objects: if manager_open {
-                Self::authored_object_count(&self.tabs)
-            } else {
-                0
-            },
-            manager_rows: &rows,
-        };
-        let ask = self.draw_chrome_pass(ctx, read, true);
-        self.apply_drawing_chrome(ask, now);
-    }
-
-    /// One split for both call sites. `floating` picks which of the surface's
-    /// two entry points runs.
-    fn draw_chrome_pass(
-        &mut self,
-        ctx: &egui::Context,
-        read: DrawingRead<'_>,
-        floating: bool,
-    ) -> crate::surfaces::drawing_chrome::DrawingChromeAsk {
-        // Split into disjoint borrows, like the surface registry above: the
-        // chrome is drawn through `&mut` while what it reads is borrowed from
-        // the rest of the application.
-        let Self {
-            surfaces,
-            tabs,
-            active_tab,
-            toolrail,
-            drawing_presets,
-            ..
-        } = self;
-        let env = drawing_env(&tabs[*active_tab], toolrail, drawing_presets, read);
-        if floating {
-            surfaces.drawing_chrome.draw_floating(ctx, &env)
-        } else {
-            surfaces.drawing_chrome.draw_pinned_panel(ctx, &env)
-        }
-    }
-
-    /// The `QUANTICK_TEXT_NOTE` hook's other half: place a note in the middle
-    /// of the window and open its editor, through the same two calls a click
-    /// makes.
-    ///
-    /// Here rather than in the surface because every line of it is the host's:
-    /// where the visible window is, what the tape last closed at, and the saved
-    /// defaults a fresh object opens with.
-    fn place_text_note(&mut self) -> bool {
-        let Some(tool) = drawings::DRAWING_TOOLS
-            .into_iter()
-            .find(|tool| tool.holds_text())
-        else {
-            return false;
-        };
-        let point = {
-            let pane = self.drawing_pane();
-            let slots = pane.slots();
-            if pane.last_chart_area.is_none() || slots == 0 {
-                // No laid-out pane yet, and nothing to place against. The ask
-                // stands and the next frame tries again.
-                return false;
-            }
-            let close = pane
-                .closed_bar(slots.saturating_sub(1))
-                .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-                .unwrap_or(1.0);
-            let centre = pane
-                .last_auto_range
-                .filter(|(lo, hi)| hi > lo)
-                .map_or(close, |(lo, hi)| (lo + hi) / 2.0);
-            let visible = DEMO_VISIBLE_SLOTS.min(slots);
-            let slot = (slots - visible / 2).min(slots.saturating_sub(1));
-            drawings::ChartPoint::at_time(slot as f32 + 0.5, centre, pane.slot_open_time(slot))
-        };
-        // Through the same door the click path uses, saved defaults and all —
-        // and on the same pane every drawing surface reads, so the index the
-        // editor opens on is the object this just placed.
-        let fresh = drawings::new_drawing_from_defaults(&self.drawing_presets, tool);
-        let placed = self.drawing_pane_mut().drawings.place_with(
-            tool,
-            &drawings::DrawingBand::Price,
-            point,
-            |_| fresh,
-        );
-        if placed && let Some(index) = self.drawing_pane().drawings.selected() {
-            self.begin_inline_text_edit(index);
-        }
-        placed
-    }
-
-    /// The selected object's screen bounding box, expanded by the anchor
-    /// radius — the rectangle the inspector must not cover. Projected on the
-    /// focused pane, which is where the selection lives.
-    fn drawing_bbox_on_screen(&self, chart: egui::Rect, index: usize) -> Option<egui::Rect> {
-        let total = self.drawing_pane().slots();
-        let auto = self.drawing_pane().last_auto_range?;
-        let scale = self.drawing_pane().price_view.scale(
-            auto,
-            self.drawing_pane().last_chart_top,
-            self.drawing_pane().last_chart_top + self.drawing_pane().last_chart_height,
-        );
-        let history_right = self
-            .drawing_pane()
-            .last_lane_divider_x
-            .unwrap_or(chart.right());
-        let drawing = self.drawing_pane().drawings.items().get(index)?;
-        let points =
-            self.drawing_pane()
-                .projected_drawing_points(drawing, history_right, total, &scale);
-        let first = points.first()?;
-        let mut bbox = egui::Rect::from_min_max(*first, *first);
-        for point in &points {
-            bbox.extend_with(*point);
-        }
-        // What the tool paints, which is not always where its anchors are: a
-        // fixed-range profile anchors at one price and covers the axis. Every
-        // popup that keeps clear of an object reads this rectangle, so asking
-        // the anchors alone is what let a panel land in the middle of a
-        // profile while believing it had walked around it.
-        let bbox = drawing.tool.painted_bounds(bbox, chart);
-        Some(bbox.expand(DRAWING_ANCHOR_RADIUS_PX))
-    }
-
-    /// Carry out what the replay interface asked for.
-    /// Whether the action reached its destination. Only a transport control
-    /// can fail to — see the drop below — and the one caller that gets a
-    /// single shot at it (the scripted seek) reads this before spending it.
-    fn apply_replay_action(&mut self, action: ReplayAction) -> bool {
-        match action {
-            ReplayAction::Open(request) => {
-                let (tab, config) = self.active_with_config();
-                tab.open_replay(config, *request);
-                true
-            }
-            ReplayAction::Close => {
-                let (tab, config) = self.active_with_config();
-                tab.close_replay(config);
-                true
-            }
-            ReplayAction::Control(control) => {
-                // A dropped transport click is not worth a retry queue: the
-                // worker drains commands every 8 ms, so a full channel means
-                // the click was already superseded.
-                if let Err(e) = self
-                    .active_tab()
-                    .commands
-                    .try_send(FeedCommand::Replay(control))
-                {
-                    tracing::debug!(
-                        target: "quantick::app",
-                        event_code = "REPLAY_COMMAND_DROPPED",
-                        reason = %e,
-                        "transport command not queued"
-                    );
-                    return false;
-                }
-                true
-            }
-        }
-    }
-}
-
-impl QuantickApp {
-    /// Where a scripted right-click should land to reach `pane`'s menu.
-    ///
-    /// Mid-height, and mid-pane horizontally, off the geometry the draw
-    /// published — so the click lands on the canvas rather than on the axis,
-    /// the legend or the divider handle. `None` until the pane has drawn once
-    /// (no divider yet), and `None` for the tape on a canvas that has none:
-    /// there is no tape menu to open where there is no tape.
-    fn scripted_context_menu_pos(&self, pane: ContextMenuPane) -> Option<egui::Pos2> {
-        let flow = &self.active_tab().flow_pane;
-        // The axis's menu lives on the gutter, off the canvas entirely — the
-        // draw publishes that band the same way it publishes the divider.
-        if pane == ContextMenuPane::Axis {
-            return Some(flow.last_price_gutter?.center());
-        }
-        // The time axis, likewise off the canvas — and its own published band,
-        // because the segment past the lane divider is the tape's.
-        if pane == ContextMenuPane::Time {
-            return Some(flow.last_time_strip?.center());
-        }
-        let rect = flow.last_chart_rect?;
-        let divider = flow.last_lane_divider_x;
-        let x = match (pane, divider) {
-            (ContextMenuPane::Tape, Some(divider)) => (divider + rect.right()) / 2.0,
-            (ContextMenuPane::Tape, None) => return None,
-            // Axis and Time returned above; anything else is the candles'
-            // canvas.
-            (_, Some(divider)) => (rect.left() + divider) / 2.0,
-            (_, None) => rect.center().x,
-        };
-        Some(egui::pos2(x, rect.center().y))
-    }
-
-    /// Where `QUANTICK_POINTER` puts the mouse this frame, in window points.
-    ///
-    /// Resolved against the *drawing* area rather than the whole chart, so a
-    /// fraction means a place among the candles whatever share of the canvas
-    /// the live lane has taken, and against the flow pane for the same reason
-    /// [`Self::scripted_context_menu_pos`] does — one canvas per capture.
-    /// `None` until the pane has drawn once: there is no candle area to be a
-    /// fraction of before then, and guessing one would park the pointer
-    /// somewhere the author did not ask for.
-    fn scripted_pointer_pos(&self) -> Option<egui::Pos2> {
-        let fraction = self.harness.pointer()?;
-        let flow = &self.active_tab().flow_pane;
-        let candles = flow.drawing_area(flow.last_chart_rect?);
-        Some(egui::pos2(
-            candles.left() + fraction.x * candles.width(),
-            candles.top() + fraction.y * candles.height(),
-        ))
-    }
-
-    /// Deliver the parked pointer, every frame it is parked.
-    fn push_scripted_pointer(&self, raw_input: &mut egui::RawInput) {
-        if let Some(position) = self.scripted_pointer_pos() {
-            raw_input.events.push(egui::Event::PointerMoved(position));
-        }
     }
 }
 
@@ -3939,366 +1800,6 @@ impl eframe::App for QuantickApp {
 }
 
 impl QuantickApp {
-    /// The scripted view hooks (`QUANTICK_CANDLE_WIDTH`, `QUANTICK_PAN_PX`),
-    /// re-applied every frame.
-    ///
-    /// Every frame rather than once at boot, for two reasons. A pan needs bars
-    /// to move over and at boot there are none — repeating it is what makes
-    /// `QUANTICK_PAN_PX=-9000` mean "as far left as it goes" whatever the
-    /// zoom: each frame pushes, the per-frame clamp holds, and the view settles
-    /// on the projection margin.
-    ///
-    /// And the view is *rebuilt* under both hooks by anything that re-cuts the
-    /// series: `ChartPane::reset_series` hands back a fresh `Viewport`, which a
-    /// replay autostart does before its first frame. A zoom set at boot was
-    /// therefore thrown away, and every scripted capture of a recorded session
-    /// photographed the default zoom rather than the one it asked for.
-    ///
-    /// A run with neither variable set does nothing here.
-    fn apply_scripted_view(&mut self) {
-        let (width, pan) = self.harness.scripted_view();
-        if width.is_none() && pan.is_none() {
-            return;
-        }
-        let pane = &mut self.active_tab_mut().flow_pane;
-        if let Some(px) = width {
-            pane.viewport.set_px_per_bar(px);
-        }
-        let slots = pane.slots();
-        if let Some(dx) = pan
-            && slots > 0
-        {
-            pane.viewport.pan_pixels(dx, slots);
-        }
-    }
-
-    /// The `QUANTICK_LOAD_OLDER` hook: press "+ older" this many times, once
-    /// the chart has something to page back from.
-    ///
-    /// Goes through [`Tab::request_older_history`] — the very function the
-    /// toolbar button calls — rather than reaching for the feed command itself,
-    /// so a run under this hook exercises the trader's path including its
-    /// loading indicator, and cannot drift from it.
-    ///
-    /// One page per frame at most: the pages are answered asynchronously and
-    /// the feed serves one request at a time, so firing them together would
-    /// have every page after the first refused and answered empty — a capture
-    /// of the drop path rather than of the feature.
-    fn apply_load_older(&mut self) {
-        let Some(pages) = self.harness.load_older_pages() else {
-            return;
-        };
-        if self.active_tab().flow_pane.slots() == 0 {
-            // Nothing charted yet. Wait, but not forever.
-            if self.harness.spend_load_older_frame().gave_up {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "LOAD_OLDER_AUTOSTART_GAVE_UP",
-                    pages,
-                    frames_waited = crate::harness::LOAD_OLDER_HOOK_FRAMES,
-                    action = "chart_left_as_it_is",
-                    "QUANTICK_LOAD_OLDER found no bars to page back from"
-                );
-            }
-            return;
-        }
-        if self.active_tab().loading.is_active(LoadingTask::History) {
-            // The previous page is still coming. Asking now would be refused
-            // and answered empty, which is not what the hook is for.
-            return;
-        }
-        let (tab, config) = self.active_with_config();
-        tab.request_older_history(config);
-        self.harness.load_older_page_sent();
-    }
-
-    /// The `QUANTICK_HISTORY_NOTE` hook: the sentence a settled reach leaves,
-    /// held up over a chart for as long as the hook's budget lasts.
-    ///
-    /// Re-applied every frame rather than raised once, the way
-    /// `QUANTICK_PAN_PX` re-applies its drag — and for a reason a one-shot
-    /// could not survive. Switching source clears the note along with the run
-    /// that raised it, exactly as it should: a new market has nothing to say
-    /// about the last one's press. But a launch under `QUANTICK_REPLAY_AUTOSTART`
-    /// *is* a source switch, arriving a second after the first bars, so a note
-    /// raised once was swept away before any shutter could open on it.
-    ///
-    /// Holding it re-raises only while it is absent, so the surface itself is
-    /// unchanged — the same sentence, from the same call, in the same lane.
-    /// When the budget runs out the note keeps its ordinary
-    /// [`crate::tab::HISTORY_NOTE_LINGER`] from the last raise and then leaves
-    /// on its own, so even a hooked run photographs a note that expires.
-    fn apply_history_note_hook(&mut self) {
-        let Some(end) = self.harness.history_note_ending() else {
-            return;
-        };
-        if self.harness.spend_history_note_frame().gave_up {
-            tracing::info!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HISTORY_NOTE_HOOK_RELEASED",
-                ending = end.action(),
-                frames_held = crate::harness::HISTORY_NOTE_HOOK_FRAMES,
-                action = "note_left_to_expire",
-                "QUANTICK_HISTORY_NOTE let go of its sentence"
-            );
-            return;
-        }
-        let tab = self.active_tab();
-        // Nothing charted yet, or the note is already up: nothing to raise.
-        //
-        // And never while a request is out. Paired with `QUANTICK_LOAD_OLDER`
-        // — which the harness table pairs it with — the press clears the note
-        // and sends a real `load_older`, and re-raising here would paint a
-        // settled verdict over a request still in flight, with the spinner
-        // turning above it. That is the dishonesty this branch removes, and a
-        // hook has no business manufacturing it for a capture.
-        if tab.flow_pane.slots() == 0
-            || tab.history_note().is_some()
-            || tab.loading.is_active(LoadingTask::History)
-        {
-            return;
-        }
-        // Always `Some`: the hook only ever holds an ending the env read above
-        // kept, and it keeps only endings that have words.
-        let Some(notice) = end.notice() else {
-            return;
-        };
-        self.active_tab_mut().raise_history_note(notice);
-    }
-
-    /// The `QUANTICK_LOAD_OLDER_CANDLES` hook: the history menu's "+ older
-    /// candles" entry, pressed without a hand, once per frame at most.
-    ///
-    /// Same shape and same reasons as [`Self::apply_load_older`], against a
-    /// different record: it goes through `Tab::request_older_ohlcv_history`
-    /// rather than the feed command, so a run under this hook exercises the
-    /// trader's own path; it waits, because there is nothing to reach back
-    /// *from* until the opening request has landed; and it gives up rather
-    /// than hanging a capture on a venue that never answers.
-    fn apply_load_older_candles(&mut self) {
-        let Some(spans) = self.harness.load_older_candle_spans() else {
-            return;
-        };
-        let capabilities = self.active_tab().capabilities(&self.config);
-        // Waiting costs budget, but a *slower* budget. A span really being
-        // fetched is the feature working, and charging it at the same rate as
-        // an empty chart would give up around the fourth of the documented
-        // thirteen spans. Charging it nothing, though, is how a venue that
-        // simply never answers hangs a capture run for the life of the
-        // process — which is the exact failure this counter exists to bound,
-        // and what the doc above promises it does. So a fetching frame spends
-        // one tick of a budget scaled to how long fetching legitimately takes.
-        if self
-            .active_tab()
-            .loading
-            .is_active(LoadingTask::VenueHistory)
-        {
-            if self.harness.spend_load_older_candles_frame().gave_up {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "LOAD_OLDER_CANDLES_AUTOSTART_GAVE_UP",
-                    spans,
-                    frames_waited = crate::harness::LOAD_OLDER_CANDLES_HOOK_FRAMES,
-                    reason = "venue_never_answered",
-                    action = "chart_left_as_it_is",
-                    "QUANTICK_LOAD_OLDER_CANDLES gave up waiting for a span to arrive"
-                );
-            }
-            return;
-        }
-        if !self.active_tab().can_load_older_candles(capabilities) {
-            // Nothing to reach back *from* yet, or the venue's record starts
-            // here. Both are worth waiting a bounded while for, and both end
-            // the same way; the log names what the tab held so an operator can
-            // tell them apart.
-            if self.harness.spend_load_older_candles_frame().gave_up {
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "LOAD_OLDER_CANDLES_AUTOSTART_GAVE_UP",
-                    spans,
-                    frames_waited = crate::harness::LOAD_OLDER_CANDLES_HOOK_FRAMES,
-                    candles_held = self.active_tab().venue_candles_held(),
-                    ohlcv_history = capabilities.ohlcv_history,
-                    action = "chart_left_as_it_is",
-                    "QUANTICK_LOAD_OLDER_CANDLES found nothing to reach back from"
-                );
-            }
-            return;
-        }
-        // Only a request that actually went out costs a *span*. A full command
-        // channel is a busy frame, not a span delivered, and counting it as one
-        // would quietly shorten the reach the operator asked for — but it still
-        // costs a frame of budget, or a permanently saturated channel leaves
-        // the hook armed for the life of the process with nothing ever logged.
-        if self
-            .active_tab_mut()
-            .request_older_ohlcv_history(capabilities)
-        {
-            self.harness.load_older_candles_span_sent();
-        } else if self.harness.spend_load_older_candles_frame().gave_up {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "LOAD_OLDER_CANDLES_AUTOSTART_GAVE_UP",
-                spans,
-                frames_waited = crate::harness::LOAD_OLDER_CANDLES_HOOK_FRAMES,
-                reason = "request_never_queued",
-                action = "chart_left_as_it_is",
-                "QUANTICK_LOAD_OLDER_CANDLES could not get a request out"
-            );
-        }
-    }
-
-    /// Play the alarm cues every tab's armed instances asked for this
-    /// frame, and empty their queues.
-    ///
-    /// Every tab, not only the active one: a tab the trader is not looking
-    /// at keeps its feed running and its instances judging, and an alarm
-    /// exists precisely to be heard when the eyes are elsewhere.
-    ///
-    /// One cue per *distinct* cue per frame, across every tab. The kernel's
-    /// repeat rule has already thinned each instance's stream to one per
-    /// bar (or one per cooldown); this is the second, blunter guard, for
-    /// the frame that ingested a burst of prints and closed several bars at
-    /// once — four identical beeps stacked into one instant are one noise,
-    /// not four alarms.
-    ///
-    /// Deduplicating by *cue* rather than collapsing to one is the whole
-    /// point of letting a preset choose a sound: a trader who gave two
-    /// regions two sounds did it to tell them apart, and swallowing the
-    /// second because it shared a frame with the first would hide a signal
-    /// and leave no trace that it had. The frame's cues go to the sink as
-    /// one batch, which plays them in order — so the second is heard after
-    /// the first rather than instead of it. The set of sounds is small and
-    /// fixed, so this is bounded by the catalogue however busy the tape.
-    ///
-    /// Per frame, but cheap: the walk is over a handful of tabs whose
-    /// queues are empty on every frame but the one a signal happened on,
-    /// and the sink is only asked when something is queued.
-    fn play_pending_alarms(&mut self) {
-        // Order of first request, duplicates dropped — `dedup` alone would
-        // only collapse neighbours.
-        let mut distinct: Vec<crate::audio::Cue> = Vec::new();
-        for tab in &mut self.tabs {
-            for cue in tab.pending_alarm_sounds.drain(..) {
-                if !distinct.contains(&cue) {
-                    distinct.push(cue);
-                }
-            }
-        }
-        if distinct.is_empty() {
-            return;
-        }
-        let outcome = self.alerts.play(&distinct);
-        self.report_alert_attempt(outcome);
-    }
-
-    /// Record whether a sound actually reached the trader.
-    ///
-    /// A notification that never arrived is reported, never assumed — so a
-    /// first failure raises a toast rather than waiting for the trader to
-    /// reopen the arming dialog, which they may never do. Only the *first*
-    /// of a run: a build with no audio backend fails on every alarm, and a
-    /// toast per bar would be its own noise. A success clears the reason, so
-    /// one transient refusal does not leave a permanent red line behind it.
-    fn report_alert_attempt(&mut self, outcome: Result<(), &'static str>) {
-        match outcome {
-            Ok(()) => self.alert_failure = None,
-            Err(reason) => {
-                let first = self.alert_failure.as_deref() != Some(reason);
-                self.alert_failure = Some(reason.to_owned());
-                if first {
-                    self.show_agent_toast(format!("no alarm sound was played: {reason}"));
-                }
-            }
-        }
-    }
-
-    /// Arm one instance on a drawing: compile the form, warm the trigger on
-    /// the bars already closed (gates shut, so nothing fires from history),
-    /// attach it, and start the paper host listening. `Err` carries the
-    /// human-readable refusal for the dialog to show.
-    /// Duplicate the selected drawing with everything riding it.
-    ///
-    /// The one door, because a duplication is not only a copied mark: an
-    /// armed strategy rides the drawing today and whatever docks next will
-    /// ride it too. Two call sites — the hotkey and the context bar — each
-    /// spelling out "copy, then carry" is a third one that copies and
-    /// forgets, and a band that silently loses its bot is exactly the class
-    /// of silence this change exists to end.
-    ///
-    /// Rate: rare — one keystroke or one click.
-    fn duplicate_selected_drawing(&mut self) {
-        let side = self.active_tab().drawing_side();
-        let Some(duplicated) = self
-            .drawing_pane_mut()
-            .drawings
-            .duplicate_selected(DUPLICATE_OFFSET_BARS)
-        else {
-            return;
-        };
-        self.carry_strategy_to_duplicate(side, duplicated);
-    }
-
-    /// Carry an armed strategy across a duplication.
-    ///
-    /// A copied region is a region the trader wants watched the same way —
-    /// duplicating the band and then re-typing the form is a step that only
-    /// exists because the copy forgot. The copy is armed through
-    /// [`Self::arm_strategy_instance`], the same door the dialog uses, from
-    /// the stored form the source kept: one construction path, so a copy
-    /// cannot quietly differ from what the dialog would have built.
-    ///
-    /// **Only a watching instance travels.** A source that is `Done`, or
-    /// disarmed for any reason, was stopped — by the trader's own hand, by
-    /// a rejected entry, by a spent one shot — and a copy that springs back
-    /// to life places orders the trader last said no to. The copy lands
-    /// offset to the *right*, the direction that makes a dead span live
-    /// again, so Ctrl+D would otherwise be the one gesture that silently
-    /// revives what was deliberately stopped.
-    ///
-    /// **State does not travel either.** The copy starts `Armed`, with a
-    /// fresh ruler warmed on this pane's own bars, a fresh alarm (no
-    /// inherited cooldown, no inherited preview mark) and no order id.
-    /// Cloning a `Fired` instance would hang a second badge on one order.
-    ///
-    /// A refusal is reported rather than swallowed: `duplicate_selected`
-    /// clones `hidden`, `off_series` and `foreign_market` verbatim — only
-    /// `locked` is reset — and arming refuses all three. A trader who
-    /// pressed hide-all and then Ctrl+D would otherwise unhide to two
-    /// identical bands wearing one badge and believe both were watching.
-    ///
-    /// Rate: rare — one keystroke.
-    fn carry_strategy_to_duplicate(
-        &mut self,
-        side: pane::PaneSide,
-        duplicated: drawings::Duplicated,
-    ) {
-        use quantick_strategy::ArmedState;
-        let Some((spec, label)) = self
-            .active_tab()
-            .pane(side)
-            .strategies
-            .for_drawing(duplicated.source)
-            .filter(|instance| {
-                matches!(
-                    instance.armed.state(),
-                    ArmedState::Armed | ArmedState::Fired { .. } | ArmedState::InPosition
-                )
-            })
-            .map(|instance| (instance.spec.clone(), instance.preset.clone()))
-        else {
-            return;
-        };
-        if let Err(reason) = self.arm_strategy_instance(side, duplicated.copy, &spec, label) {
-            self.note_workspace(format!("the copy carries no strategy: {reason}"));
-        }
-    }
-
     fn arm_strategy_instance(
         &mut self,
         side: pane::PaneSide,
@@ -4401,938 +1902,6 @@ impl QuantickApp {
         }
         Ok(())
     }
-
-    /// The `QUANTICK_REPLAY_RESTART_AFTER` hook: press the transport's own
-    /// Restart once the session has closed that many round trips.
-    ///
-    /// The seek is the only way to put a closed trade ahead of the tape the
-    /// chart holds — the recording starts over, the round trips stay in the
-    /// ledger because they happened, and their fills are now at instants no
-    /// bar on screen covers. That is the state the marks used to stack on
-    /// the edge bar in, and it takes a click on a transport button a
-    /// scripted capture cannot make. Nothing happens without a recording
-    /// playing: there is no timeline to seek on a live feed.
-    ///
-    /// Consumed once, whether or not the trades ever arrived — an env var
-    /// is a request for this run, not a standing rule.
-    fn apply_replay_restart(&mut self) {
-        let Some(after) = self.harness.replay_restart_after() else {
-            return;
-        };
-        let tab = self.active_tab();
-        if tab.replay.is_none() || tab.paper.session_trades().len() < after {
-            return;
-        }
-        // Spent only once the transport took it. A hook that cleared itself
-        // on a dropped command would leave the capture photographing an
-        // un-seeked timeline while the harness believed otherwise; the next
-        // frame simply tries again.
-        if self.apply_replay_action(ReplayAction::Control(ReplayControl::Restart)) {
-            self.harness.replay_restart_taken();
-        }
-    }
-}
-
-impl QuantickApp {
-    /// One frame of the application: drain, lay out the chrome, draw the
-    /// chart.
-    ///
-    /// Everything `update` does that is not eframe's own bookkeeping, so a
-    /// test can run a real frame against a headless [`egui::Context`] and read
-    /// what was painted — the only honest way to assert that a chart is on
-    /// screen rather than a blank rectangle.
-    fn draw_frame(&mut self, ctx: &egui::Context, now: Instant) {
-        if let Some(last) = self.last_frame {
-            self.frames.record((now - last).as_secs_f32() * 1000.0);
-        }
-        self.last_frame = Some(now);
-
-        self.drain_tabs();
-        // A "load older" outcome is a passing remark: it leaves after
-        // `tab::HISTORY_NOTE_LINGER` whether or not anyone read it. Every tab,
-        // not only the one on screen — a background tab keeps draining, so it
-        // can settle a run while hidden, and bringing it forward minutes later
-        // must not surface a sentence about a press that is long over.
-        for tab in &mut self.tabs {
-            tab.expire_history_note(now);
-        }
-        // After the expiry, never before it: the hook re-raises a note it
-        // finds absent, and running it first would let a note expire *after*
-        // it looked, drawing one frame with an empty lane before the next
-        // raise. A shutter timed on the linger catches exactly that frame.
-        self.apply_history_note_hook();
-        if self.pending_control_access_enable {
-            self.pending_control_access_enable = false;
-            if let Some(access) = self.control_access.as_mut() {
-                access.enable(ctx);
-            }
-        }
-        // Replay determinism: a session with a control trace beside it
-        // re-injects its actions at their logical time, connected or not.
-        // Before the hook's mark, so a loaded sidecar has seeded the trace
-        // sequence the mark will take.
-        if let Some(mut access) = self.control_access.take() {
-            access.service_replay_trace(self);
-            self.control_access = Some(access);
-        }
-        if let Some(note) = self.pending_control_mark.take() {
-            let note = (!note.is_empty()).then_some(note);
-            self.take_mark(note);
-        }
-        self.apply_control_annotate_hooks();
-        // After the annotate hooks and before the gateway's own drain: a
-        // bundle captured from a launch then describes the window an
-        // assistant has already written on, which is the state a validation
-        // run is actually asking about.
-        self.apply_control_evidence_hook(ctx);
-        if self
-            .control_access
-            .as_ref()
-            .is_some_and(crate::control::ControlAccess::needs_frame_service)
-            && let Some(mut access) = self.control_access.take()
-        {
-            access.begin_frame(self, ctx);
-            self.control_access = Some(access);
-        }
-        self.apply_scripted_view();
-        self.apply_drawing_demo();
-        self.apply_load_older();
-        self.apply_load_older_candles();
-        self.apply_drawing_draft();
-        self.apply_venue_history_demo();
-        self.apply_frvp_demo();
-        self.apply_avwap_demo();
-        self.apply_strategy_demo();
-        self.apply_replay_restart();
-        self.apply_maximize_hook(ctx);
-        self.maybe_emit_summary(now, ctx);
-        self.maintain_workspace(ctx);
-
-        let bg = pane::background_color(&self.style);
-        // Rail shortcuts first: Esc/1/2 must be read before any widget can
-        // claim the keyboard this frame.
-        self.toolrail.handle_keys(ctx);
-        self.handle_tab_keys(ctx);
-        self.handle_drawing_keys(ctx, now);
-        // Chrome panels claim their zones outside-in (§5): menu and toolbar
-        // on top, the status line at the very bottom with the replay
-        // transport directly above it, then the edge-docked drawing rail and
-        // the right dock. The chart keeps whatever remains.
-        self.draw_menu_bar(ctx);
-        if let Some(access) = self.control_access.as_mut() {
-            access.draw_panel(ctx);
-        }
-        self.draw_toolbar(ctx);
-        // Before the dialog is drawn, so a double click on a pane or a curve
-        // opens it on the same frame the gesture happened rather than the next.
-        self.open_requested_indicator_settings();
-        self.draw_indicator_settings(ctx);
-        self.draw_indicator_legends(ctx);
-        // **After** the dialogs above, and that placement is load-bearing.
-        // The preview watermark reads whether a settings dialog is previewing
-        // an unapplied draft, and `draw_indicator_settings` is what sets that
-        // — so an environment built before it would put the banner on screen
-        // a frame after the legend chip that says the same thing, and take it
-        // off a frame later too. Two surfaces the trader reads as one is this
-        // repo's own bug class; sixteen milliseconds of it is still one
-        // frame a capture can photograph.
-        //
-        // It is also where the windows this pass drew used to sit: the
-        // appearance and footprint panels ran after the toolbar that toggles
-        // them, so a click on LOOK opens the panel on the same frame rather
-        // than the next.
-        // A pane's right-click asked to arm one of its drawings. Drained
-        // here, into the surface that owns the dialog: the click happens
-        // while the canvas draws, which is later in this frame than the
-        // surfaces are, so the dialog opens on the next one — a frame the
-        // trader cannot see, and the price of the dialog no longer living in
-        // the trunk.
-        let sides: SmallVec<[pane::PaneSide; MAX_CANVAS_PANES]> =
-            self.active_tab().sides().collect();
-        for side in sides {
-            let request = self
-                .active_tab_mut()
-                .pane_mut(side)
-                .strategy_popup_request
-                .take();
-            if let Some(drawing) = request {
-                let form = crate::strategy_presets::StoredPreset::starting_point(
-                    quantick_engine::Side::Buy,
-                );
-                let tab = self.active_tab().id;
-                self.surfaces.strategy_popup.open(tab, side, drawing, form);
-            }
-        }
-        // The bar rules the arming dialog's alarm section reads: a share of
-        // the bar only means something where the rule closes on a count.
-        // Built only while that dialog is open, like the open markets below.
-        // `hooks_pending` is the frame a capture hook opens a surface from
-        // inside `draw_all`: it is not open yet when this runs, but it is
-        // about to be, and it must not draw its first frame against an empty
-        // environment.
-        let staging = self.surfaces.hooks_pending();
-        let counted_bar_sides: SmallVec<[pane::PaneSide; MAX_CANVAS_PANES]> =
-            if staging || self.surfaces.strategy_popup.is_open() {
-                let tab = self.active_tab();
-                tab.sides()
-                    .filter(|side| tab.pane(*side).state.progress().is_some())
-                    .collect()
-            } else {
-                SmallVec::new()
-            };
-        // The markets tabs are showing: the dialog greys out removing one of
-        // those, because a tab left on a symbol the catalog no longer offers
-        // gets silently retargeted by the next SOURCE correction. Built only
-        // while the dialog is open — it is a `String` pair per tab, and no
-        // frame should pay for it to be thrown away.
-        let open_markets: Vec<(String, String)> =
-            if staging || self.surfaces.source_picker.is_open() {
-                self.tabs
-                    .iter()
-                    .map(|tab| (tab.feed_id.clone(), tab.symbol.clone()))
-                    .collect()
-            } else {
-                Vec::new()
-            };
-        // Split into disjoint borrows: the surfaces are drawn through `&mut`
-        // while the environment they read is borrowed from the rest of the
-        // application. That the compiler insists on the split is the port
-        // working — a surface cannot be handed the trunk it is being kept
-        // out of.
-        let Self {
-            surfaces: registry,
-            workspace,
-            style,
-            footprint_config,
-            tabs,
-            active_tab,
-            indicator_settings,
-            indicator_settings_target,
-            config,
-            added_symbols,
-            alert_failure,
-            ..
-        } = self;
-        let focused_tab = &tabs[*active_tab];
-        // Read once. `focused_pane` resolves the same side internally, and
-        // the answer is not a field lookup — it reads the layout, because
-        // focus on a collapsed pane is focus on nothing.
-        let focused_side = focused_tab.focused_side();
-        let focused_pane = focused_tab.pane(focused_side);
-        let surfaces = registry.draw_all(
-            ctx,
-            &crate::surfaces::SurfaceEnv {
-                bookmarks: workspace.session().bookmarks(),
-                now,
-                indicator_preview_area: indicator_preview_area(
-                    tabs,
-                    indicator_settings.as_ref(),
-                    *indicator_settings_target,
-                ),
-                focused_chart_area: focused_pane.last_chart_area,
-                style,
-                footprint: focused_pane.footprint_config(footprint_config),
-                footprint_customized: focused_pane.footprint_override.is_some(),
-                focused_side,
-                config,
-                added_symbols,
-                open_markets: &open_markets,
-                active_tab: focused_tab.id,
-                counted_bar_sides: &counted_bar_sides,
-                alert_failure: alert_failure.as_deref(),
-            },
-        );
-        if let Some(name) = surfaces.save_workspace_as {
-            self.save_named_workspace(&name);
-        }
-        if let Some(style) = surfaces.style {
-            self.style = style;
-            self.style_revision = self.style_revision.saturating_add(1);
-        }
-        // After the assignment, never before: the log line reports the
-        // appearance that is now in force, and the revision it landed on.
-        if let Some(request) = surfaces.log_style_change {
-            self.emit_style_changed(request.applied_preset);
-        }
-        // The audition goes through the one speaker every armed instance
-        // shares, and reports a sound that could not be heard exactly as a
-        // missed signal would.
-        if let Some(cue) = surfaces.test_alert {
-            let outcome = self.alerts.play(&[cue]);
-            self.report_alert_attempt(outcome);
-        }
-        if let Some(request) = surfaces.arm_strategy {
-            let outcome = self.arm_strategy_instance(
-                request.side,
-                request.drawing,
-                &request.form,
-                request.label,
-            );
-            self.surfaces.strategy_popup.settle_arm(outcome);
-        }
-        if let Some(request) = surfaces.market {
-            self.apply_market_request(request);
-        }
-        if let Some(change) = surfaces.footprint {
-            self.apply_footprint_change(change);
-        }
-        if surfaces.undo_drawing {
-            let pane = self.drawing_pane_mut();
-            pane.drawings.undo();
-            // Same orphan risk as the keyboard undo: the drawing an armed
-            // instance rides may just have been taken away.
-            pane.sweep_strategy_orphans();
-        }
-        self.poll_script_files();
-        self.maintain_indicator_state();
-        self.maintain_chart_layers();
-        // This tab's judgement about its own feed, taken once for the frame:
-        // the status bar reads it here and the corner reads it below, and two
-        // readings a millisecond apart could disagree about whether a budget
-        // had run out.
-        let stall = self
-            .active_tab()
-            .stall_at(&self.config, metrics::wall_clock_ms());
-        let offline_accent = self.feed_offline_accent(stall.as_ref());
-        let status = self.status_model();
-        let status_response = statusbar::draw(ctx, &status, &mut self.tz, offline_accent);
-        if status_response.open_trading_tab {
-            self.dock.open_tab(DockTab::Trading);
-        }
-        // Above the status bar, below the canvas: the layout tabs.
-        self.draw_layout_strip(ctx);
-        self.draw_layout_delete_confirm(ctx);
-        // The browser window and, while the *active* tab plays a session, its
-        // transport bar. A background tab's recording keeps advancing on its
-        // own feed thread; what it does not get is the strip, which speaks for
-        // one tab at a time (§11).
-        let replay_action = {
-            let Self {
-                replay_view,
-                tabs,
-                active_tab,
-                config,
-                ..
-            } = self;
-            let tab = &tabs[*active_tab];
-            // The instruments the download tab offers with one click. A dated
-            // contract rolls every couple of months, and typing `WINV26` from
-            // memory is not a thing a trader should have to get right to see
-            // what they can replay.
-            //
-            // Filtered by what the download source actually serves, which the
-            // source itself answers: offering a Binance pair to a MetaTrader
-            // exporter would be a click that can only end in a refusal, and
-            // the chart behind this window is often on another venue entirely.
-            let serves = replay_view.download_provider();
-            let market = crate::replay_view::MarketMenu {
-                current: (config.provider_of(&tab.feed_id) == Some(serves))
-                    .then_some(tab.symbol.as_str()),
-                catalogue: config
-                    .feeds
-                    .iter()
-                    .filter(|feed| feed.provider == serves)
-                    .flat_map(|feed| feed.symbols.iter().map(String::as_str))
-                    .collect(),
-            };
-            replay_view.draw(ctx, tab.replay.as_ref(), &market)
-        };
-        if let Some(action) = replay_action {
-            self.apply_replay_action(action);
-        }
-        // A folder the trader just pointed the browser at is written down on
-        // the frame they pointed it, not at exit: "it forgot my folder again"
-        // must not be one crash away.
-        if let Some(pick) = self.replay_view.take_folder_change() {
-            self.write_replay_folder(pick.as_deref());
-        }
-        // The same, for the tick that decides whether yesterday is on the
-        // chart. Either row can have been the one clicked; the browser owns
-        // the setting, so there is one place to pick the change up.
-        if let Some(enabled) = self.replay_view.take_day_before_change() {
-            self.write_replay_day_before(enabled);
-        }
-        {
-            // The focused pane's objects: the toolbox lists and manages what a
-            // click on the canvas would act on.
-            let side = self.active_tab().focused_side();
-            // The flag lives with the window it opens, so it travels through
-            // a local rather than a `&mut` handed out of the surface.
-            let mut manager_open = self.surfaces.drawing_chrome.manager_open();
-            {
-                let Self {
-                    toolrail,
-                    tabs,
-                    active_tab,
-                    ..
-                } = self;
-                let tab = &mut tabs[*active_tab];
-                toolrail.draw(ctx, &mut tab.pane_mut(side).drawings, &mut manager_open);
-            }
-            self.surfaces.drawing_chrome.set_manager_open(manager_open);
-        }
-        // A star clicked this frame is on disk this frame, like the replay
-        // folder above: the pinned rail is what the trader reaches for without
-        // looking, and rebuilding it after a crash is not a thing anyone
-        // should have to do twice.
-        if self.toolrail.take_favorites_change() {
-            self.write_favorites();
-        }
-        let dock_response = {
-            let Self {
-                dock,
-                tabs,
-                active_tab,
-                replay_view,
-                tz,
-                ..
-            } = self;
-            // The Trading tab speaks for the market on screen: one tab, one
-            // simulator, and the dock reads the active tab's — exactly like
-            // the tape and the session panel beside it.
-            let Tab {
-                flow_pane,
-                replay,
-                paper,
-                ..
-            } = &mut tabs[*active_tab];
-            let orderflow = flow_pane
-                .orderflow
-                .as_mut()
-                .expect("the flow pane is built with a tape and never drops it");
-            dock.draw(
-                ctx,
-                &mut DockEnv {
-                    orderflow,
-                    replay_view,
-                    replay: replay.as_ref(),
-                    paper,
-                    tz: *tz,
-                },
-            )
-        };
-        // The strategy editor is a window of the active tab's ticket, drawn
-        // whatever the dock is showing and whether it is showing at all: it
-        // is opened from the Trading tab but it does not belong to it, and a
-        // trader who opens it and then looks at the ledger has not asked for
-        // it to close.
-        if self.active_tab_mut().paper.draw_strategy_editor(ctx) {
-            self.persist_order_strategies();
-        }
-        if dock_response.restart_book_capture {
-            self.active_tab_mut().restart_book_capture();
-        }
-        if let Some(action) = dock_response.replay_action {
-            // A click that lost its slot has the trader's next click behind
-            // it; only the one-shot hook below cares about the answer.
-            let _ = self.apply_replay_action(action);
-        }
-        // The ledger's jump-to-trade: center the flow pane on the round
-        // trip's midpoint, the object manager's own "select and centre".
-        //
-        // The covering lookup, the same one the marks are painted through:
-        // a trade the flow chart's bars do not reach has nowhere to be
-        // centred on, and scrolling to the clamped edge instead would land
-        // the trader on a bar holding no mark and no explanation. Saying so
-        // is the whole of the handling — the row stays in the ledger.
-        //
-        // The message names the flow chart rather than "the chart": in a
-        // split tab the time pane keeps its own, longer window, so the same
-        // round trip can be off this one and painted on that one.
-        if let Some((opened, closed)) = dock_response.navigate_to_trade {
-            let tab = self.active_tab_mut();
-            let covered = tab
-                .flow_pane
-                .covering_slot_at_time(opened)
-                .zip(tab.flow_pane.covering_slot_at_time(closed));
-            match covered {
-                Some((entry, exit)) => {
-                    let pane = &mut tab.flow_pane;
-                    if let Some(area) = pane.last_chart_area {
-                        let slots = pane.slots();
-                        let mid = (entry + exit) as f32 / 2.0;
-                        pane.viewport.center_on_bar(mid, area.width(), slots);
-                    }
-                }
-                None => {
-                    // Said as an event as well as on screen: an operator
-                    // driving the ledger without eyes on the toast must be
-                    // able to tell a refusal from a silent no-op.
-                    tracing::info!(
-                        target: "quantick::app",
-                        event_code = "TRADE_NAVIGATE_OFF_TAPE",
-                        opened_ms = opened,
-                        closed_ms = closed,
-                        "jump-to-trade refused: the flow chart has no bar for the fills"
-                    );
-                    tab.paper.show_toast(
-                        "This trade is outside the bars on the flow chart - nothing to centre on."
-                            .to_owned(),
-                    );
-                }
-            }
-        }
-        if dock_response.pick_trades_dir {
-            self.open_trades_dir_picker();
-        }
-        if dock_response.order_strategies_changed {
-            self.persist_order_strategies();
-        }
-        if dock_response.cmd_trading_changed {
-            self.persist_cmd_trading();
-        }
-        if dock_response.risk_settings_changed {
-            self.persist_risk_settings();
-        }
-        self.poll_trades_dir_picker();
-        self.poll_workspace_picker();
-        // The pinned inspector is chrome: declared before the central canvas
-        // so the chart pays its width, exactly like the dock.
-        self.draw_pinned_inspector(ctx, now);
-        // Respawn the feed if the feed/symbol selection changed (resets the
-        // chart), then apply any bar-type change (no-op if unchanged).
-        let (tab, config) = self.active_with_config();
-        tab.maybe_switch_feed(config);
-        // Both deferrals settle here, a frame after the click that armed
-        // them, so the frame carrying the change paints its overlay first.
-        let Self {
-            tabs,
-            config,
-            style,
-            pane_ids,
-            ..
-        } = self;
-        for tab in tabs.iter_mut() {
-            tab.apply_pending_layout(config, style, pane_ids);
-        }
-        // Right after panes appear and markets switch, so a pane built this
-        // frame is seeded this frame and a tab that changed symbol swaps its
-        // drawings before anything paints them.
-        self.maintain_layouts();
-        self.active_tab_mut().apply_spec_changes();
-        // Waits owned by other components, mirrored level-style each frame so
-        // the overlay needs no push notifications from either.
-        let replay_loading = self.replay_view.is_loading();
-        let book_syncing = self.active_tab().tape().is_syncing();
-        let tab = self.active_tab_mut();
-        tab.loading
-            .set_active(LoadingTask::ReplaySession, replay_loading);
-        tab.loading.set_active(LoadingTask::BookSync, book_syncing);
-
-        let mut notice_action = feed_notice::NoticeAction::None;
-        // Read before the canvas borrows `self`, and answered after it lets go.
-        let popup_tab = self.active_tab().id;
-        let popup_open = self.feed_popup_tab == Some(popup_tab);
-        let mut chip_clicked = false;
-        let mut dismissed = false;
-        // Where the corner landed, and so whether there was one at all. A feed
-        // that recovered while the popup was open closes it, rather than
-        // leaving a stale explanation over a chart that is fine again.
-        let mut chip_rect = None;
-        // The layer menu offers what this source can produce; resolved once
-        // here rather than per pane, per entry, inside the canvas.
-        let capabilities = self.active_tab().capabilities(&self.config);
-        // Same one-per-frame resolution for the side-honesty label the
-        // footprint legend carries.
-        let side_inferred = self.active_tab().side_note(&self.config).is_some();
-        // Told before the canvas paints, not after: the object holding the
-        // words the editor is showing must stand down on the *same* frame,
-        // or the note flashes its placeholder under the field for one.
-        self.sync_content_editing();
-        // Raised by a placement that wants its note typed, and handed to the
-        // chrome below: the flag belongs to the editor that owns the caret,
-        // not to the canvas that asks for it.
-        let mut begin_text_edit = false;
-        egui::CentralPanel::default()
-            .frame(egui::Frame::none().fill(bg))
-            .show(ctx, |ui| {
-                let area = ui.available_rect_before_wrap();
-                {
-                    let Self {
-                        tabs,
-                        active_tab,
-                        toolrail,
-                        drawing_presets,
-                        style,
-                        tz,
-                        layer_actions,
-                        footprint_config,
-                        ..
-                    } = self;
-                    let mut chrome = CanvasChrome {
-                        toolrail,
-                        presets: drawing_presets,
-                        begin_text_edit: &mut begin_text_edit,
-                        style,
-                        tz: *tz,
-                        capabilities,
-                        side_inferred,
-                        footprint: footprint_config,
-                        layers: layer_actions,
-                    };
-                    tabs[*active_tab].draw_canvas(ui, area, &mut chrome);
-                }
-                // The grid and the indicator state belong to the window, not
-                // to the pane whose menu switched them.
-                self.apply_layer_actions();
-                let tab = self.active_tab();
-                // Each wait on the surface it is about. The panes published
-                // their rects on the draw just above, so these are this
-                // frame's geometry rather than the previous one's.
-                let history_note = tab.history_note();
-                loading::overlay_scoped(ui, area, &tab.loading, LoadingScope::Whole, history_note);
-                // A scope whose surface is not on screen falls back to the
-                // canvas rather than dropping its wait: the flow pane is not
-                // painted in the Time layout, and a flow-only layout has no
-                // time pane, and in both cases the wait is still running. A
-                // spinner in the wrong place is a placement complaint; a
-                // missing one reads as a frozen application.
-                loading::overlay_scoped(
-                    ui,
-                    tab.flow_pane.last_area.unwrap_or(area),
-                    &tab.loading,
-                    LoadingScope::Flow,
-                    history_note,
-                );
-                let time_panes: Vec<egui::Rect> = tab
-                    .panes()
-                    .filter(|(_, side)| matches!(side, PaneSide::Time(_)))
-                    .filter_map(|(pane, _)| pane.last_area)
-                    .collect();
-                if time_panes.is_empty() {
-                    loading::overlay_scoped(
-                        ui,
-                        area,
-                        &tab.loading,
-                        LoadingScope::TimePanes,
-                        history_note,
-                    );
-                } else {
-                    for rect in time_panes {
-                        loading::overlay_scoped(
-                            ui,
-                            rect,
-                            &tab.loading,
-                            LoadingScope::TimePanes,
-                            history_note,
-                        );
-                    }
-                }
-                // And the feed's own report, in the corner rather than over
-                // the chart. Progress never gets here: a first connection and a
-                // history block already have the loading overlay above, and a
-                // second badge beside it would be the interface talking about
-                // itself twice.
-                if let Some(report) = feed_notice::report(&tab.notice, stall.as_ref())
-                    && report.is_offline()
-                {
-                    // Measured once, then handed to everything that needs
-                    // it: the chip's own hit test, the popup's anchor, the
-                    // dismissal test, and the scene's bounds.
-                    let chip = feed_notice::chip_rect(ui.painter(), area);
-                    chip_rect = Some(chip);
-                    chip_clicked = feed_notice::draw_chip(ui, chip, &report, popup_open);
-                    // A pane with nothing on it has room to say why, and a
-                    // corner chip alone on a blank canvas is a puzzle. One
-                    // muted line, no border and no buttons — the way out is
-                    // still the corner.
-                    //
-                    // Not while the popup is up. The line and the popup carry
-                    // the same headline, and on the empty chart that is
-                    // exactly where both of them draw: one sentence, twice, a
-                    // hand apart.
-                    if !popup_open && let Some((pane_rect, 0)) = tab.starved_pane() {
-                        feed_notice::draw_empty_pane_note(ui.painter(), pane_rect, &report);
-                    }
-                    if popup_open {
-                        // A click anywhere else puts it away, measured against
-                        // the rectangles that were actually drawn — so a click
-                        // on the edge of what the trader can see is never read
-                        // as a click outside it, and the popup is laid out
-                        // once rather than measured again to ask.
-                        let popup;
-                        (notice_action, popup) = feed_notice::draw_popup(ui, area, chip, &report);
-                        dismissed = ui.input(|input| {
-                            input.pointer.any_click()
-                                && input
-                                    .pointer
-                                    .interact_pos()
-                                    .is_some_and(|at| !popup.contains(at) && !chip.contains(at))
-                        });
-                    }
-                }
-            });
-        if begin_text_edit {
-            self.surfaces.drawing_chrome.request_text_edit();
-        }
-        // Floating drawing controls must be registered after the opaque
-        // central canvas so they stay in front of the chart. That is why the
-        // drawing chrome is the one surface `Surfaces::draw_all` does not
-        // draw: it is anchored *to* the chart rather than floating over the
-        // window, so it is commanded by name from here instead.
-        self.draw_drawing_chrome(ctx, now);
-        // The menus above may have disarmed a bot over a resting retest
-        // limit; its cancel goes to the simulator on this same frame, not
-        // on the next print. Every tab, not just the active one: a menu
-        // click and a tab switch can land on the same frame, and the old
-        // tab's feed keeps running — its cancel must not sit stranded
-        // until the tab is looked at again.
-        for tab in &mut self.tabs {
-            tab.apply_strategy_cleanup();
-        }
-        self.play_pending_alarms();
-        // Both are window chrome reading the active tab, like the offline
-        // corner and the transport strip: they speak for one market at a time.
-        let tz = self.tz;
-        self.active_tab_mut().paper.draw_report_window(ctx, tz);
-        self.settle_paper_panels(now);
-        // Both controls go through the tab's own methods, which are also what
-        // the registered control-plane actions call: a click and a named call
-        // must be able to disagree about nothing.
-        match notice_action {
-            feed_notice::NoticeAction::None => {}
-            feed_notice::NoticeAction::Reconnect => {
-                let (tab, config) = self.active_with_config();
-                let _ = tab.reconnect_feed(config);
-            }
-            feed_notice::NoticeAction::Reload => {
-                let (tab, config) = self.active_with_config();
-                let _ = tab.reload_feed(config);
-            }
-        }
-        self.feed_chip_rect = chip_rect;
-        self.feed_popup_tab = feed_notice::popup_still_open(
-            popup_open,
-            chip_clicked,
-            chip_rect.is_some(),
-            dismissed,
-            notice_action,
-        )
-        .then_some(popup_tab);
-        // Live feed: keep polling the channel ~60×/s without busy-spinning.
-        ctx.request_repaint_after(Duration::from_millis(16));
-    }
-
-    /// Every tab takes in what its feed sent this frame, on screen or not.
-    ///
-    /// §11: switching tabs never tears a feed down, so a background tab has to
-    /// keep draining — its channels are bounded, and one left full backs its
-    /// feed thread up until the market it is showing is hours behind. The
-    /// indicator workers are fed on the same pass, so a tab brought forward is
-    /// already current rather than rebuilding on the frame it appears.
-    fn drain_tabs(&mut self) {
-        let config = &self.config;
-        let progressive_history = self.progressive_history;
-        let history_reach = self.history_reach;
-        let history_reach_span_minutes = self.history_reach_span_minutes;
-        let venue_lead_in = self.venue_lead_in;
-        let mut trades = 0_u64;
-        for tab in &mut self.tabs {
-            let before = tab.live_trades;
-            tab.drain_feed();
-            for pane in tab.panes_mut() {
-                pane.apply_indicator_events();
-            }
-            tab.drain_book_feed();
-            tab.drain_notices();
-            // Heartbeat for the recorder. The lifecycle calls elsewhere already
-            // start it at every point that knows the market changed; this one
-            // makes "always recording" true by construction, so a start command
-            // lost to a momentarily full channel heals on the next frame
-            // instead of leaving the session silently unrecorded. Free while it
-            // is running: one bool read and an early return.
-            tab.ensure_book_capture(config);
-            // MetaTrader narrows its capabilities when the bridge says hello,
-            // after the pane may already have asked and been told there was
-            // nothing held. Watching the edge is what asks again once the
-            // answer can be a real one.
-            // The switch lives on the window, the request is phrased by the
-            // tab: mirrored here so every tab asks the way the trader last
-            // said, including one opened after the choice was made.
-            tab.progressive_history = progressive_history;
-            tab.history_reach = history_reach;
-            tab.history_reach_span_minutes = history_reach_span_minutes;
-            // Through the setter, not the field: flipping the lead-in refolds
-            // the prefix, and a tab that only had the field written would keep
-            // drawing the answer to the previous choice until the next candle
-            // landed. Idempotent, so the steady state costs one comparison.
-            tab.set_venue_lead_in(venue_lead_in);
-            tab.poll_ohlcv_capability(config);
-            trades += tab.live_trades - before;
-        }
-        // What the window ingested, across every market it is holding.
-        self.trades_since_summary += trades;
-    }
-
-    /// Tab shortcuts (§10): `Ctrl+T` new, `Ctrl+W` close, `Ctrl+Tab` cycle.
-    fn handle_tab_keys(&mut self, ctx: &egui::Context) {
-        // Focus-gated like `handle_drawing_keys` (audit MINOR-13): typing in
-        // the source picker's field with Ctrl held must never close the tab
-        // under it — closing is instant and currently irreversible.
-        if ctx.memory(|memory| memory.focused().is_some()) {
-            return;
-        }
-        let (new_tab, close_tab, next, previous) = ctx.input_mut(|input| {
-            (
-                input.consume_shortcut(&NEW_TAB_SHORTCUT),
-                input.consume_shortcut(&CLOSE_TAB_SHORTCUT),
-                input.consume_shortcut(&NEXT_TAB_SHORTCUT),
-                input.consume_shortcut(&PREVIOUS_TAB_SHORTCUT),
-            )
-        });
-        if new_tab {
-            self.surfaces.source_picker.open(&self.config);
-        }
-        if close_tab {
-            self.close_tab(self.active_tab);
-        }
-        if next {
-            self.cycle_tab(1);
-        }
-        if previous {
-            self.cycle_tab(-1);
-        }
-    }
-
-    /// Do what the "Open market" dialog settled on.
-    fn apply_market_request(&mut self, request: crate::surfaces::MarketRequest) {
-        use crate::surfaces::MarketRequest;
-        match request {
-            MarketRequest::Open { feed_id, symbol } => self.open_tab(feed_id, symbol, None),
-            MarketRequest::Add { feed_id, symbol } => match self.add_symbol(&feed_id, &symbol) {
-                Ok(()) => {
-                    self.surfaces.source_picker.close();
-                    self.open_tab(feed_id, symbol, None);
-                }
-                // The dialog stays open carrying the reason: the user is one
-                // keystroke from a symbol that does fit, and closing would
-                // make the refusal look like a crash.
-                Err(reason) => self.surfaces.source_picker.refuse(reason),
-            },
-            MarketRequest::Remove { feed_id, symbol } => self.remove_symbol(&feed_id, &symbol),
-        }
-    }
-
-    /// Put `symbol` in feed `feed_id`'s catalog and remember it across
-    /// restarts. Reports whether the catalog took it.
-    ///
-    /// The config file itself is never written: it is hand-written, comments
-    /// and all, and a program that rewrote it would eat them. The addition
-    /// lives in its own sidecar, which the next launch folds back in before
-    /// the config is validated (see [`crate::symbols_file`]).
-    fn add_symbol(&mut self, feed_id: &str, symbol: &str) -> Result<(), String> {
-        // Against the *whole* config, on a copy. A symbol is not just a name
-        // in a list: it takes part in every cross-check the config has, and
-        // the MetaTrader port map is one where a single mapped symbol offered
-        // by two feeds is a configuration the app refuses to load. Persisting
-        // one of those would write a file that kills the next launch — and the
-        // error would name the config, which is not the file that broke.
-        let mut candidate = self.config.clone();
-        if !candidate.add_symbol(feed_id, symbol) {
-            return Err(format!(
-                "{} already offers {symbol}",
-                self.config.feed_name(feed_id)
-            ));
-        }
-        candidate.validate()?;
-        self.config = candidate;
-        self.added_symbols.add(feed_id, symbol);
-        if let Err(error) = symbols_file::save(self.workspace.symbols_path(), &self.added_symbols) {
-            // The catalog took it for this session either way; what is lost is
-            // the next launch, and the user is told which file did not take it.
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "SYMBOL_CATALOG_WRITE_FAILED",
-                path = %self.workspace.symbols_path().display(),
-                error = %error,
-                action = "addition_is_session_only",
-                "cannot write the added-symbols file"
-            );
-        }
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "SYMBOL_ADDED",
-            feed = %feed_id,
-            symbol = %symbol,
-            path = %self.workspace.symbols_path().display(),
-            action = "open_in_new_tab",
-            "a symbol was added from the source picker"
-        );
-        Ok(())
-    }
-
-    /// Take a user-added `symbol` back out of feed `feed_id`'s catalog.
-    ///
-    /// Only ever a catalog edit: a tab already showing that market keeps
-    /// streaming it. The picker will not offer this for a market a tab is on,
-    /// which is what stops the selection correction from retargeting it.
-    fn remove_symbol(&mut self, feed_id: &str, symbol: &str) {
-        if !self.config.remove_symbol(feed_id, symbol) {
-            return;
-        }
-        self.added_symbols.remove(feed_id, symbol);
-        if let Err(error) = symbols_file::save(self.workspace.symbols_path(), &self.added_symbols) {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "SYMBOL_CATALOG_WRITE_FAILED",
-                path = %self.workspace.symbols_path().display(),
-                error = %error,
-                action = "removal_is_session_only",
-                "cannot write the added-symbols file"
-            );
-        }
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "SYMBOL_REMOVED",
-            feed = %feed_id,
-            symbol = %symbol,
-            path = %self.workspace.symbols_path().display(),
-            action = "leave_open_tabs_alone",
-            "a user-added symbol left the catalog"
-        );
-    }
-
-    /// Carry out what the tab strip asked for.
-    fn apply_tab_action(&mut self, action: TabAction) {
-        match action {
-            TabAction::Activate(index) => {
-                if index < self.tabs.len() {
-                    self.active_tab = index;
-                }
-            }
-            TabAction::Close(index) => self.close_tab(index),
-            TabAction::New => self.surfaces.source_picker.open(&self.config),
-        }
-    }
-}
-
-/// The interval a saved bar rule names, when it is a time rule at all — a
-/// workspace that recorded `tick:50` for a context chart is a file written by
-/// hand, and the chart opens on the default rather than on a guess.
-fn saved_time_interval(text: Option<&str>) -> Option<i64> {
-    text.and_then(|text| match BarSpec::parse(text) {
-        Ok(BarSpec::Time(ms)) => Some(ms),
-        _ => None,
-    })
-}
-
-/// Every context chart's opening interval, top to bottom, from the rules a
-/// workspace saved. A rule that is not a time rule keeps the default for its
-/// slot so the slots after it still line up with their charts. A file written
-/// before the stack existed carries only `time_bars`, which is the top chart's.
-fn saved_context_intervals(bars: &[String], time_bars: Option<&str>) -> Vec<i64> {
-    if bars.is_empty() {
-        return saved_time_interval(time_bars).into_iter().collect();
-    }
-    bars.iter()
-        .map(|text| {
-            saved_time_interval(Some(text)).unwrap_or(crate::time_header::DEFAULT_INTERVAL_MS)
-        })
-        .collect()
 }
 
 crate::hooks::declare_hooks![
